Create a recount3 RangedSummarizedExperiment gene or exon object

Once you have identified a project you want to work with, you can use this function to construct a recount3 RangedSummarizedExperiment-class (RSE) object at the gene or exon expression feature level. This function will retrieve the data, cache it, then assemble the RSE object.

create_rse(
  project_info,
  type = c("gene", "exon", "jxn"),
  annotation = annotation_options(project_info$organism),
  bfc = BiocFileCache::BiocFileCache(),
  jxn_format = c("ALL", "UNIQUE"),
  recount3_url = "http://idies.jhu.edu/recount3/data"
)

Arguments

project_info
A data.frame() with one row that contains the information for the project you are interested in. You can find which project to work on using todo().

type
A character(1) specifying whether you want to access gene, exon, or exon-exon junction counts.

annotation
A character(1) specifying which annotation you want to download. Only used when type is either gene or exon.

bfc
A BiocFileCache-class object where the files will be cached to.

jxn_format
A character(1) specifying whether the exon-exon junction files are derived from all the reads (ALL) or only the uniquely mapping read counts (UNIQUE). Note that UNIQUE is only available for some projects.

recount3_url
A character(1) specifying the home URL for recount3 or a local directory where you have mirrored recount3.

Value

A RangedSummarizedExperiment-class object.

Examples

## Find the project you are interested in
## (temporarily build this data.frame)
proj_info <- data.frame(
    project = "ERP110066",
    project_home = "data_sources/sra",
    organism = "human"
)

## Create a RSE object at the gene level
rse_gene_ERP110066 <- create_rse(proj_info)

#> 2020-09-27 21:45:40 downloading and reading the metadata.

#> using temporary cache /tmp/RtmpjEQHEA/BiocFileCache

#> 2020-09-27 21:45:40 caching file sra.sra.ERP110066.MD.gz.

#> adding rname 'http://idies.jhu.edu/recount3/data/human/data_sources/sra/metadata/66/ERP110066/sra.sra.ERP110066.MD.gz'

#> 2020-09-27 21:45:40 caching file sra.recount_project.ERP110066.MD.gz.

#> adding rname 'http://idies.jhu.edu/recount3/data/human/data_sources/sra/metadata/66/ERP110066/sra.recount_project.ERP110066.MD.gz'

#> 2020-09-27 21:45:41 caching file sra.recount_qc.ERP110066.MD.gz.

#> adding rname 'http://idies.jhu.edu/recount3/data/human/data_sources/sra/metadata/66/ERP110066/sra.recount_qc.ERP110066.MD.gz'

#> 2020-09-27 21:45:41 caching file sra.recount_seq_qc.ERP110066.MD.gz.

#> adding rname 'http://idies.jhu.edu/recount3/data/human/data_sources/sra/metadata/66/ERP110066/sra.recount_seq_qc.ERP110066.MD.gz'

#> 2020-09-27 21:45:42 caching file sra.recount_pred.ERP110066.MD.gz.

#> adding rname 'http://idies.jhu.edu/recount3/data/human/data_sources/sra/metadata/66/ERP110066/sra.recount_pred.ERP110066.MD.gz'

#> Warning: Replacing study_acc by study in sra.sra.ERP110066.MD.gz

#> Warning: Replacing run_acc by external_id in sra.sra.ERP110066.MD.gz

#> Warning: Replacing study_acc by study in sra.recount_qc.ERP110066.MD.gz

#> Warning: Replacing run_acc by external_id in sra.recount_qc.ERP110066.MD.gz

#> 2020-09-27 21:45:43 downloading and reading the feature information.

#> 2020-09-27 21:45:43 caching file human.gene_sums.G026.gtf.gz.

#> adding rname 'http://idies.jhu.edu/recount3/data/human/annotations/gene_sums/human.gene_sums.G026.gtf.gz'

#> 2020-09-27 21:45:44 downloading and reading the counts: 1024 samples across 63856 features.

#> 2020-09-27 21:45:44 caching file sra.gene_sums.ERP110066.G026.gz.

#> adding rname 'http://idies.jhu.edu/recount3/data/human/data_sources/sra/gene_sums/66/ERP110066/sra.gene_sums.ERP110066.G026.gz'

#> 2020-09-27 21:45:52 construcing the RangedSummarizedExperiment (rse) object.

## Explore the resulting RSE gene object
rse_gene_ERP110066

#> class: RangedSummarizedExperiment 
#> dim: 63856 1024 
#> metadata(9): time_created recount3_version ... jxn_format recount3_url
#> assays(1): counts
#> rownames(63856): ENSG00000278704.1 ENSG00000277400.1 ...
#>   ENSG00000182484.15_PAR_Y ENSG00000227159.8_PAR_Y
#> rowData names(10): source type ... havana_gene tag
#> colnames(1024): ERR2713604 ERR2713605 ... ERR2713998 ERR2713999
#> colData names(173): rail_id external_id ... recount_pred.pred.type
#>   BigWigURL

## Information about how this RSE object was made
metadata(rse_gene_ERP110066)

#> $time_created
#> [1] "2020-09-27 21:45:52 UTC"
#> 
#> $recount3_version
#>           package ondiskversion loadedversion
#> recount3 recount3        0.99.0        0.99.0
#>                                                 path
#> recount3 /usr/local/lib/R/host-site-library/recount3
#>                                           loadedpath attached is_base
#> recount3 /usr/local/lib/R/host-site-library/recount3     TRUE   FALSE
#>                date       source md5ok                            library
#> recount3 2020-09-27 Bioconductor    NA /usr/local/lib/R/host-site-library
#> 
#> $project
#> [1] "ERP110066"
#> 
#> $project_home
#> [1] "data_sources/sra"
#> 
#> $type
#> [1] "gene"
#> 
#> $organism
#> [1] "human"
#> 
#> $annotation
#> [1] "gencode_v26"
#> 
#> $jxn_format
#> [1] "ALL"
#> 
#> $recount3_url
#> [1] "http://idies.jhu.edu/recount3/data"
#> 

## Number of genes by number of samples
dim(rse_gene_ERP110066)

#> [1] 63856  1024

## Information about the genes
rowRanges(rse_gene_ERP110066)

#> GRanges object with 63856 ranges and 10 metadata columns:
#>                              seqnames            ranges strand |   source
#>                                 <Rle>         <IRanges>  <Rle> | <factor>
#>          ENSG00000278704.1 GL000009.2       56140-58376      - |  ENSEMBL
#>          ENSG00000277400.1 GL000194.1      53590-115018      - |  ENSEMBL
#>          ENSG00000274847.1 GL000194.1      53594-115055      - |  ENSEMBL
#>          ENSG00000277428.1 GL000195.1       37434-37534      - |  ENSEMBL
#>          ENSG00000276256.1 GL000195.1       42939-49164      - |  ENSEMBL
#>                        ...        ...               ...    ... .      ...
#>   ENSG00000124334.17_PAR_Y       chrY 57184101-57197337      + |   HAVANA
#>   ENSG00000185203.12_PAR_Y       chrY 57201143-57203357      - |   HAVANA
#>    ENSG00000270726.6_PAR_Y       chrY 57190738-57208756      + |   HAVANA
#>   ENSG00000182484.15_PAR_Y       chrY 57207346-57212230      + |   HAVANA
#>    ENSG00000227159.8_PAR_Y       chrY 57212184-57214397      - |   HAVANA
#>                                type     score     phase                gene_id
#>                            <factor> <numeric> <integer>            <character>
#>          ENSG00000278704.1     gene      2237      <NA>      ENSG00000278704.1
#>          ENSG00000277400.1     gene      2179      <NA>      ENSG00000277400.1
#>          ENSG00000274847.1     gene      1599      <NA>      ENSG00000274847.1
#>          ENSG00000277428.1     gene       101      <NA>      ENSG00000277428.1
#>          ENSG00000276256.1     gene      2195      <NA>      ENSG00000276256.1
#>                        ...      ...       ...       ...                    ...
#>   ENSG00000124334.17_PAR_Y     gene      2504      <NA> ENSG00000124334.17_P..
#>   ENSG00000185203.12_PAR_Y     gene      1054      <NA> ENSG00000185203.12_P..
#>    ENSG00000270726.6_PAR_Y     gene       773      <NA> ENSG00000270726.6_PA..
#>   ENSG00000182484.15_PAR_Y     gene      4618      <NA> ENSG00000182484.15_P..
#>    ENSG00000227159.8_PAR_Y     gene      1306      <NA> ENSG00000227159.8_PA..
#>                                         gene_type   gene_name       level
#>                                       <character> <character> <character>
#>          ENSG00000278704.1         protein_coding  BX004987.1           3
#>          ENSG00000277400.1         protein_coding  AC145212.2           3
#>          ENSG00000274847.1         protein_coding  AC145212.1           3
#>          ENSG00000277428.1               misc_RNA       Y_RNA           3
#>          ENSG00000276256.1         protein_coding  AC011043.1           3
#>                        ...                    ...         ...         ...
#>   ENSG00000124334.17_PAR_Y         protein_coding        IL9R           2
#>   ENSG00000185203.12_PAR_Y              antisense      WASIR1           2
#>    ENSG00000270726.6_PAR_Y   processed_transcript AJ271736.10           2
#>   ENSG00000182484.15_PAR_Y transcribed_unproces..      WASH6P           2
#>    ENSG00000227159.8_PAR_Y unprocessed_pseudogene    DDX11L16           2
#>                                     havana_gene         tag
#>                                     <character> <character>
#>          ENSG00000278704.1                 <NA>        <NA>
#>          ENSG00000277400.1                 <NA>        <NA>
#>          ENSG00000274847.1                 <NA>        <NA>
#>          ENSG00000277428.1                 <NA>        <NA>
#>          ENSG00000276256.1                 <NA>        <NA>
#>                        ...                  ...         ...
#>   ENSG00000124334.17_PAR_Y OTTHUMG00000022720.1         PAR
#>   ENSG00000185203.12_PAR_Y OTTHUMG00000022676.3         PAR
#>    ENSG00000270726.6_PAR_Y OTTHUMG00000184987.2         PAR
#>   ENSG00000182484.15_PAR_Y OTTHUMG00000022677.5         PAR
#>    ENSG00000227159.8_PAR_Y OTTHUMG00000022678.1         PAR
#>   -------
#>   seqinfo: 374 sequences from an unspecified genome; no seqlengths

## Sample metadata
colnames(colData(rse_gene_ERP110066))

#>   [1] "rail_id"                                                           
#>   [2] "external_id"                                                       
#>   [3] "study"                                                             
#>   [4] "sra.sample_acc.x"                                                  
#>   [5] "sra.experiment_acc"                                                
#>   [6] "sra.submission_acc"                                                
#>   [7] "sra.submission_center"                                             
#>   [8] "sra.submission_lab"                                                
#>   [9] "sra.study_title"                                                   
#>  [10] "sra.study_abstract"                                                
#>  [11] "sra.study_description"                                             
#>  [12] "sra.experiment_title"                                              
#>  [13] "sra.design_description"                                            
#>  [14] "sra.sample_description"                                            
#>  [15] "sra.library_name"                                                  
#>  [16] "sra.library_strategy"                                              
#>  [17] "sra.library_source"                                                
#>  [18] "sra.library_selection"                                             
#>  [19] "sra.library_layout"                                                
#>  [20] "sra.paired_nominal_length"                                         
#>  [21] "sra.paired_nominal_stdev"                                          
#>  [22] "sra.library_construction_protocol"                                 
#>  [23] "sra.platform_model"                                                
#>  [24] "sra.sample_attributes"                                             
#>  [25] "sra.experiment_attributes"                                         
#>  [26] "sra.spot_length"                                                   
#>  [27] "sra.sample_name"                                                   
#>  [28] "sra.sample_title"                                                  
#>  [29] "sra.sample_bases"                                                  
#>  [30] "sra.sample_spots"                                                  
#>  [31] "sra.run_published"                                                 
#>  [32] "sra.size"                                                          
#>  [33] "sra.run_total_bases"                                               
#>  [34] "sra.run_total_spots"                                               
#>  [35] "sra.num_reads"                                                     
#>  [36] "sra.num_spots"                                                     
#>  [37] "sra.read_info"                                                     
#>  [38] "sra.run_alias"                                                     
#>  [39] "sra.run_center_name"                                               
#>  [40] "sra.run_broker_name"                                               
#>  [41] "sra.run_center"                                                    
#>  [42] "recount_project.project"                                           
#>  [43] "recount_project.organism"                                          
#>  [44] "recount_project.file_source"                                       
#>  [45] "recount_project.metadata_source"                                   
#>  [46] "recount_project.date_processed"                                    
#>  [47] "recount_qc.aligned_reads%.chrm"                                    
#>  [48] "recount_qc.aligned_reads%.chrx"                                    
#>  [49] "recount_qc.aligned_reads%.chry"                                    
#>  [50] "recount_qc.bc_auc.all_reads_all_bases"                             
#>  [51] "recount_qc.bc_auc.all_reads_annotated_bases"                       
#>  [52] "recount_qc.bc_auc.unique_reads_all_bases"                          
#>  [53] "recount_qc.bc_auc.unique_reads_annotated_bases"                    
#>  [54] "recount_qc.bc_auc.all_%"                                           
#>  [55] "recount_qc.bc_auc.unique_%"                                        
#>  [56] "recount_qc.bc_frag.count"                                          
#>  [57] "recount_qc.bc_frag.kallisto_count"                                 
#>  [58] "recount_qc.bc_frag.kallisto_mean_length"                           
#>  [59] "recount_qc.bc_frag.mean_length"                                    
#>  [60] "recount_qc.bc_frag.mode_length"                                    
#>  [61] "recount_qc.bc_frag.mode_length_count"                              
#>  [62] "recount_qc.exon_fc.all_%"                                          
#>  [63] "recount_qc.exon_fc.unique_%"                                       
#>  [64] "recount_qc.exon_fc_count_all.total"                                
#>  [65] "recount_qc.exon_fc_count_all.assigned"                             
#>  [66] "recount_qc.exon_fc_count_unique.total"                             
#>  [67] "recount_qc.exon_fc_count_unique.assigned"                          
#>  [68] "recount_qc.gene_fc.all_%"                                          
#>  [69] "recount_qc.gene_fc.unique_%"                                       
#>  [70] "recount_qc.gene_fc_count_all.total"                                
#>  [71] "recount_qc.gene_fc_count_all.assigned"                             
#>  [72] "recount_qc.gene_fc_count_unique.total"                             
#>  [73] "recount_qc.gene_fc_count_unique.assigned"                          
#>  [74] "recount_qc.intron_sum"                                             
#>  [75] "recount_qc.intron_sum_%"                                           
#>  [76] "recount_qc.star.%_of_chimeric_reads"                               
#>  [77] "recount_qc.star.%_of_chimeric_reads2"                              
#>  [78] "recount_qc.star.%_of_reads_mapped_to_multiple_loci"                
#>  [79] "recount_qc.star.%_of_reads_mapped_to_multiple_loci2"               
#>  [80] "recount_qc.star.%_of_reads_mapped_to_too_many_loci"                
#>  [81] "recount_qc.star.%_of_reads_mapped_to_too_many_loci2"               
#>  [82] "recount_qc.star.%_of_reads_unmapped:_other"                        
#>  [83] "recount_qc.star.%_of_reads_unmapped:_other2"                       
#>  [84] "recount_qc.star.%_of_reads_unmapped:_too_many_mismatches"          
#>  [85] "recount_qc.star.%_of_reads_unmapped:_too_many_mismatches2"         
#>  [86] "recount_qc.star.%_of_reads_unmapped:_too_short"                    
#>  [87] "recount_qc.star.%_of_reads_unmapped:_too_short2"                   
#>  [88] "recount_qc.star.all_mapped_reads"                                  
#>  [89] "recount_qc.star.all_mapped_reads2"                                 
#>  [90] "recount_qc.star.average_input_read_length"                         
#>  [91] "recount_qc.star.average_input_read_length2"                        
#>  [92] "recount_qc.star.average_mapped_length"                             
#>  [93] "recount_qc.star.average_mapped_length2"                            
#>  [94] "recount_qc.star.deletion_average_length"                           
#>  [95] "recount_qc.star.deletion_average_length2"                          
#>  [96] "recount_qc.star.deletion_rate_per_base"                            
#>  [97] "recount_qc.star.deletion_rate_per_base2"                           
#>  [98] "recount_qc.star.insertion_average_length"                          
#>  [99] "recount_qc.star.insertion_average_length2"                         
#> [100] "recount_qc.star.insertion_rate_per_base"                           
#> [101] "recount_qc.star.insertion_rate_per_base2"                          
#> [102] "recount_qc.star.mapping_speed,_million_of_reads_per_hour"          
#> [103] "recount_qc.star.mapping_speed,_million_of_reads_per_hour2"         
#> [104] "recount_qc.star.mismatch_rate_per_base,_%"                         
#> [105] "recount_qc.star.mismatch_rate_per_base,_%2"                        
#> [106] "recount_qc.star.number_of_chimeric_reads"                          
#> [107] "recount_qc.star.number_of_chimeric_reads2"                         
#> [108] "recount_qc.star.number_of_input_reads"                             
#> [109] "recount_qc.star.number_of_input_reads2"                            
#> [110] "recount_qc.star.number_of_reads_mapped_to_multiple_loci"           
#> [111] "recount_qc.star.number_of_reads_mapped_to_multiple_loci2"          
#> [112] "recount_qc.star.number_of_reads_mapped_to_too_many_loci"           
#> [113] "recount_qc.star.number_of_reads_mapped_to_too_many_loci2"          
#> [114] "recount_qc.star.number_of_reads_unmapped:_other"                   
#> [115] "recount_qc.star.number_of_reads_unmapped:_other2"                  
#> [116] "recount_qc.star.number_of_reads_unmapped:_too_many_mismatches"     
#> [117] "recount_qc.star.number_of_reads_unmapped:_too_many_mismatches2"    
#> [118] "recount_qc.star.number_of_reads_unmapped:_too_short"               
#> [119] "recount_qc.star.number_of_reads_unmapped:_too_short2"              
#> [120] "recount_qc.star.number_of_splices:_at/ac"                          
#> [121] "recount_qc.star.number_of_splices:_at/ac2"                         
#> [122] "recount_qc.star.number_of_splices:_annotated_(sjdb)"               
#> [123] "recount_qc.star.number_of_splices:_annotated_(sjdb)2"              
#> [124] "recount_qc.star.number_of_splices:_gc/ag"                          
#> [125] "recount_qc.star.number_of_splices:_gc/ag2"                         
#> [126] "recount_qc.star.number_of_splices:_gt/ag"                          
#> [127] "recount_qc.star.number_of_splices:_gt/ag2"                         
#> [128] "recount_qc.star.number_of_splices:_non-canonical"                  
#> [129] "recount_qc.star.number_of_splices:_non-canonical2"                 
#> [130] "recount_qc.star.number_of_splices:_total"                          
#> [131] "recount_qc.star.number_of_splices:_total2"                         
#> [132] "recount_qc.star.uniquely_mapped_reads_%"                           
#> [133] "recount_qc.star.uniquely_mapped_reads_%2"                          
#> [134] "recount_qc.star.uniquely_mapped_reads_number"                      
#> [135] "recount_qc.star.uniquely_mapped_reads_number2"                     
#> [136] "recount_qc.junction_count"                                         
#> [137] "recount_qc.junction_coverage"                                      
#> [138] "recount_qc.junction_avg_coverage"                                  
#> [139] "recount_qc.star.number_of_input_reads_both"                        
#> [140] "recount_qc.star.all_mapped_reads_both"                             
#> [141] "recount_qc.star.number_of_chimeric_reads_both"                     
#> [142] "recount_qc.star.number_of_reads_mapped_to_multiple_loci_both"      
#> [143] "recount_qc.star.number_of_reads_mapped_to_too_many_loci_both"      
#> [144] "recount_qc.star.number_of_reads_unmapped:_other_both"              
#> [145] "recount_qc.star.number_of_reads_unmapped:_too_many_mismatches_both"
#> [146] "recount_qc.star.number_of_reads_unmapped:_too_short_both"          
#> [147] "recount_qc.star.uniquely_mapped_reads_number_both"                 
#> [148] "recount_qc.star.%_mapped_reads_both"                               
#> [149] "recount_qc.star.%_chimeric_reads_both"                             
#> [150] "recount_qc.star.%_reads_mapped_to_multiple_loci_both"              
#> [151] "recount_qc.star.%_reads_mapped_to_too_many_loci_both"              
#> [152] "recount_qc.star.%_reads_unmapped:_other_both"                      
#> [153] "recount_qc.star.%_reads_unmapped:_too_many_mismatches_both"        
#> [154] "recount_qc.star.%_reads_unmapped:_too_short_both"                  
#> [155] "recount_qc.star.uniquely_mapped_reads_%_both"                      
#> [156] "recount_seq_qc.min_len"                                            
#> [157] "recount_seq_qc.max_len"                                            
#> [158] "recount_seq_qc.avg_len"                                            
#> [159] "recount_seq_qc.#distinct_quality_values"                           
#> [160] "recount_seq_qc.#bases"                                             
#> [161] "recount_seq_qc.%a"                                                 
#> [162] "recount_seq_qc.%c"                                                 
#> [163] "recount_seq_qc.%g"                                                 
#> [164] "recount_seq_qc.%t"                                                 
#> [165] "recount_seq_qc.%n"                                                 
#> [166] "recount_seq_qc.avgq"                                               
#> [167] "recount_seq_qc.errq"                                               
#> [168] "recount_pred.sample_acc.y"                                         
#> [169] "recount_pred.curated.type"                                         
#> [170] "recount_pred.curated.tissue"                                       
#> [171] "recount_pred.pattern.predict.type"                                 
#> [172] "recount_pred.pred.type"                                            
#> [173] "BigWigURL"                                                         

## Check how much memory this RSE object uses
pryr::object_size(rse_gene_ERP110066)

#> Registered S3 method overwritten by 'pryr':
#>   method      from
#>   print.bytes Rcpp

#> 285 MB

## Create a second RSE object using another annotation
rse_gene_ERP110066_gencode_v29 <- create_rse(
    proj_info,
    annotation = "gencode_v29"
)

#> 2020-09-27 21:45:52 downloading and reading the metadata.

#> 2020-09-27 21:45:52 caching file sra.sra.ERP110066.MD.gz.

#> 2020-09-27 21:45:52 caching file sra.recount_project.ERP110066.MD.gz.

#> 2020-09-27 21:45:52 caching file sra.recount_qc.ERP110066.MD.gz.

#> 2020-09-27 21:45:53 caching file sra.recount_seq_qc.ERP110066.MD.gz.

#> 2020-09-27 21:45:53 caching file sra.recount_pred.ERP110066.MD.gz.

#> Warning: Replacing study_acc by study in sra.sra.ERP110066.MD.gz

#> Warning: Replacing run_acc by external_id in sra.sra.ERP110066.MD.gz

#> Warning: Replacing study_acc by study in sra.recount_qc.ERP110066.MD.gz

#> Warning: Replacing run_acc by external_id in sra.recount_qc.ERP110066.MD.gz

#> 2020-09-27 21:45:53 downloading and reading the feature information.

#> 2020-09-27 21:45:53 caching file human.gene_sums.G029.gtf.gz.

#> adding rname 'http://idies.jhu.edu/recount3/data/human/annotations/gene_sums/human.gene_sums.G029.gtf.gz'

#> 2020-09-27 21:45:54 downloading and reading the counts: 1024 samples across 64837 features.

#> 2020-09-27 21:45:55 caching file sra.gene_sums.ERP110066.G029.gz.

#> adding rname 'http://idies.jhu.edu/recount3/data/human/data_sources/sra/gene_sums/66/ERP110066/sra.gene_sums.ERP110066.G029.gz'

#> 2020-09-27 21:46:00 construcing the RangedSummarizedExperiment (rse) object.

rowRanges(rse_gene_ERP110066_gencode_v29)

#> GRanges object with 64837 ranges and 10 metadata columns:
#>                              seqnames            ranges strand |   source
#>                                 <Rle>         <IRanges>  <Rle> | <factor>
#>          ENSG00000278704.1 GL000009.2       56140-58376      - |  ENSEMBL
#>          ENSG00000277400.1 GL000194.1      53590-115018      - |  ENSEMBL
#>          ENSG00000274847.1 GL000194.1      53594-115055      - |  ENSEMBL
#>          ENSG00000277428.1 GL000195.1       37434-37534      - |  ENSEMBL
#>          ENSG00000276256.1 GL000195.1       42939-49164      - |  ENSEMBL
#>                        ...        ...               ...    ... .      ...
#>   ENSG00000124334.17_PAR_Y       chrY 57184101-57197337      + |   HAVANA
#>   ENSG00000185203.12_PAR_Y       chrY 57201143-57203357      - |   HAVANA
#>    ENSG00000270726.6_PAR_Y       chrY 57190738-57208756      + |   HAVANA
#>   ENSG00000182484.15_PAR_Y       chrY 57207346-57212230      + |   HAVANA
#>    ENSG00000227159.8_PAR_Y       chrY 57212184-57214397      - |   HAVANA
#>                                type     score     phase                gene_id
#>                            <factor> <numeric> <integer>            <character>
#>          ENSG00000278704.1     gene      2237      <NA>      ENSG00000278704.1
#>          ENSG00000277400.1     gene      2179      <NA>      ENSG00000277400.1
#>          ENSG00000274847.1     gene      1599      <NA>      ENSG00000274847.1
#>          ENSG00000277428.1     gene       101      <NA>      ENSG00000277428.1
#>          ENSG00000276256.1     gene      2195      <NA>      ENSG00000276256.1
#>                        ...      ...       ...       ...                    ...
#>   ENSG00000124334.17_PAR_Y     gene      2504      <NA> ENSG00000124334.17_P..
#>   ENSG00000185203.12_PAR_Y     gene      1054      <NA> ENSG00000185203.12_P..
#>    ENSG00000270726.6_PAR_Y     gene       773      <NA> ENSG00000270726.6_PA..
#>   ENSG00000182484.15_PAR_Y     gene      4618      <NA> ENSG00000182484.15_P..
#>    ENSG00000227159.8_PAR_Y     gene      1306      <NA> ENSG00000227159.8_PA..
#>                                         gene_type   gene_name       level
#>                                       <character> <character> <character>
#>          ENSG00000278704.1         protein_coding  BX004987.1           3
#>          ENSG00000277400.1         protein_coding  AC145212.1           3
#>          ENSG00000274847.1         protein_coding       MAFIP           3
#>          ENSG00000277428.1               misc_RNA     RF00019           3
#>          ENSG00000276256.1         protein_coding  AC011043.1           3
#>                        ...                    ...         ...         ...
#>   ENSG00000124334.17_PAR_Y         protein_coding        IL9R           2
#>   ENSG00000185203.12_PAR_Y              antisense      WASIR1           2
#>    ENSG00000270726.6_PAR_Y   processed_transcript  AJ271736.1           2
#>   ENSG00000182484.15_PAR_Y transcribed_unproces..      WASH6P           2
#>    ENSG00000227159.8_PAR_Y unprocessed_pseudogene    DDX11L16           2
#>                                     havana_gene         tag
#>                                     <character> <character>
#>          ENSG00000278704.1                 <NA>        <NA>
#>          ENSG00000277400.1                 <NA>        <NA>
#>          ENSG00000274847.1                 <NA>        <NA>
#>          ENSG00000277428.1                 <NA>        <NA>
#>          ENSG00000276256.1                 <NA>        <NA>
#>                        ...                  ...         ...
#>   ENSG00000124334.17_PAR_Y OTTHUMG00000022720.1         PAR
#>   ENSG00000185203.12_PAR_Y OTTHUMG00000022676.3         PAR
#>    ENSG00000270726.6_PAR_Y OTTHUMG00000184987.2         PAR
#>   ENSG00000182484.15_PAR_Y OTTHUMG00000022677.5         PAR
#>    ENSG00000227159.8_PAR_Y OTTHUMG00000022678.1         PAR
#>   -------
#>   seqinfo: 406 sequences from an unspecified genome; no seqlengths

## Doesn't work right now
# rse_gene_ERP110066_refseq <- create_rse(
#     proj_info,
#     annotation = "refseq"
# )

## Create a RSE object at the exon level
if (FALSE) {
rse_exon_ERP110066 <- create_rse(
    proj_info,
    type = "exon"
)

## Explore the resulting RSE exon object
rse_exon_ERP110066

dim(rse_exon_ERP110066)
rowRanges(rse_exon_ERP110066)
pryr::object_size(rse_exon_ERP110066)
# 5.73 GB
}

## Create a RSE object at the exon-exon junction level
rse_jxn_ERP110066 <- create_rse(
    proj_info,
    type = "jxn"
)

#> 2020-09-27 21:46:00 downloading and reading the metadata.

#> 2020-09-27 21:46:00 caching file sra.sra.ERP110066.MD.gz.

#> 2020-09-27 21:46:00 caching file sra.recount_project.ERP110066.MD.gz.

#> 2020-09-27 21:46:00 caching file sra.recount_qc.ERP110066.MD.gz.

#> 2020-09-27 21:46:00 caching file sra.recount_seq_qc.ERP110066.MD.gz.

#> 2020-09-27 21:46:01 caching file sra.recount_pred.ERP110066.MD.gz.

#> Warning: Replacing study_acc by study in sra.sra.ERP110066.MD.gz

#> Warning: Replacing run_acc by external_id in sra.sra.ERP110066.MD.gz

#> Warning: Replacing study_acc by study in sra.recount_qc.ERP110066.MD.gz

#> Warning: Replacing run_acc by external_id in sra.recount_qc.ERP110066.MD.gz

#> 2020-09-27 21:46:01 downloading and reading the feature information.

#> 2020-09-27 21:46:01 caching file sra.junctions.ERP110066.ALL.RR.gz.

#> adding rname 'http://idies.jhu.edu/recount3/data/human/data_sources/sra/junctions/66/ERP110066/sra.junctions.ERP110066.ALL.RR.gz'

#> 2020-09-27 21:46:07 downloading and reading the counts: 1024 samples across 585502 features.

#> 2020-09-27 21:46:07 caching file sra.junctions.ERP110066.ALL.MM.gz.

#> adding rname 'http://idies.jhu.edu/recount3/data/human/data_sources/sra/junctions/66/ERP110066/sra.junctions.ERP110066.ALL.MM.gz'

#> 2020-09-27 21:46:29 matching exon-exon junction counts with the metadata.

#> 2020-09-27 21:46:29 caching file sra.junctions.ERP110066.ALL.ID.gz.

#> adding rname 'http://idies.jhu.edu/recount3/data/human/data_sources/sra/junctions/66/ERP110066/sra.junctions.ERP110066.ALL.ID.gz'

#> 2020-09-27 21:46:32 construcing the RangedSummarizedExperiment (rse) object.

## Explore the resulting RSE exon object
rse_jxn_ERP110066

#> class: RangedSummarizedExperiment 
#> dim: 585502 1024 
#> metadata(9): time_created recount3_version ... jxn_format recount3_url
#> assays(1): counts
#> rownames(585502): ERCC-00002:73-127:+ ERCC-00002:94-540:* ...
#>   chrY_KI270740v1_random:19966-33629:+
#>   chrY_KI270740v1_random:35065-35314:+
#> rowData names(6): length annotated ... left_annotated right_annotated
#> colnames(1024): ERR2713604 ERR2713605 ... ERR2713998 ERR2713999
#> colData names(173): rail_id external_id ... recount_pred.pred.type
#>   BigWigURL

dim(rse_jxn_ERP110066)

#> [1] 585502   1024

rowRanges(rse_jxn_ERP110066)

#> GRanges object with 585502 ranges and 6 metadata columns:
#>                                                      seqnames      ranges
#>                                                         <Rle>   <IRanges>
#>                    ERCC-00002:73-127:+             ERCC-00002      73-127
#>                    ERCC-00002:94-540:*             ERCC-00002      94-540
#>                   ERCC-00002:134-550:+             ERCC-00002     134-550
#>                   ERCC-00002:154-454:-             ERCC-00002     154-454
#>                   ERCC-00002:160-217:+             ERCC-00002     160-217
#>                                    ...                    ...         ...
#>   chrY_KI270740v1_random:19966-28241:+ chrY_KI270740v1_random 19966-28241
#>   chrY_KI270740v1_random:19966-30368:+ chrY_KI270740v1_random 19966-30368
#>   chrY_KI270740v1_random:19966-31875:+ chrY_KI270740v1_random 19966-31875
#>   chrY_KI270740v1_random:19966-33629:+ chrY_KI270740v1_random 19966-33629
#>   chrY_KI270740v1_random:35065-35314:+ chrY_KI270740v1_random 35065-35314
#>                                        strand |    length annotated  left_motif
#>                                         <Rle> | <integer> <integer> <character>
#>                    ERCC-00002:73-127:+      + |        55         0          GT
#>                    ERCC-00002:94-540:*      * |       447         0          AG
#>                   ERCC-00002:134-550:+      + |       417         0          GT
#>                   ERCC-00002:154-454:-      - |       301         0          CT
#>                   ERCC-00002:160-217:+      + |        58         0          GT
#>                                    ...    ... .       ...       ...         ...
#>   chrY_KI270740v1_random:19966-28241:+      + |      8276         0          GT
#>   chrY_KI270740v1_random:19966-30368:+      + |     10403         0          GT
#>   chrY_KI270740v1_random:19966-31875:+      + |     11910         0          GT
#>   chrY_KI270740v1_random:19966-33629:+      + |     13664         0          GT
#>   chrY_KI270740v1_random:35065-35314:+      + |       250         0          GT
#>                                        right_motif left_annotated
#>                                        <character>    <character>
#>                    ERCC-00002:73-127:+          AG              0
#>                    ERCC-00002:94-540:*          GC              0
#>                   ERCC-00002:134-550:+          AG              0
#>                   ERCC-00002:154-454:-          AC              0
#>                   ERCC-00002:160-217:+          AG              0
#>                                    ...         ...            ...
#>   chrY_KI270740v1_random:19966-28241:+          AG              0
#>   chrY_KI270740v1_random:19966-30368:+          AG              0
#>   chrY_KI270740v1_random:19966-31875:+          AG              0
#>   chrY_KI270740v1_random:19966-33629:+          AG              0
#>   chrY_KI270740v1_random:35065-35314:+          AG              0
#>                                        right_annotated
#>                                            <character>
#>                    ERCC-00002:73-127:+               0
#>                    ERCC-00002:94-540:*               0
#>                   ERCC-00002:134-550:+               0
#>                   ERCC-00002:154-454:-               0
#>                   ERCC-00002:160-217:+               0
#>                                    ...             ...
#>   chrY_KI270740v1_random:19966-28241:+               0
#>   chrY_KI270740v1_random:19966-30368:+               0
#>   chrY_KI270740v1_random:19966-31875:+               0
#>   chrY_KI270740v1_random:19966-33629:+               0
#>   chrY_KI270740v1_random:35065-35314:+               0
#>   -------
#>   seqinfo: 110 sequences from an unspecified genome; no seqlengths

pryr::object_size(rse_jxn_ERP110066)

#> 635 MB

## Remove the example rse_jxn for memory purposes
rm(rse_jxn_ERP110066)