Expand SRA attributes

This function expands the SRA attributes stored in sra.sample_attributes variable in the colData() slot of a RangedSummarizedExperiment-class produced by create_rse().

expand_sra_attributes(rse)

Arguments

rse

A RangedSummarizedExperiment-class object created by create_rse() or create_rse_manual().

Value

A RangedSummarizedExperiment-class

object with expanded metadata columns.

Details

Note that this function will work on projects available from SRA only. Furthermore, SRA attributes are project-specific. Thus, if you use this function in more than one RSE object, you won't be able to combine them easily with cbind() and will need to manually merge the colData() slots from your set of RSE files before being able to run cbind().

Author

Andrew E Jaffe modified by Leonardo Collado-Torres.

Examples

## Find all available human projects
human_projects <- available_projects()
#> 2022-11-01 21:18:42 caching file sra.recount_project.MD.gz.
#> 2022-11-01 21:18:42 caching file gtex.recount_project.MD.gz.
#> 2022-11-01 21:18:43 caching file tcga.recount_project.MD.gz.

## Find the project you are interested in
proj_info <- subset(
    human_projects,
    project == "SRP009615" & project_type == "data_sources"
)

## Create a RSE object at the gene level
rse_gene_SRP009615 <- create_rse(proj_info)
#> 2022-11-01 21:18:45 downloading and reading the metadata.
#> 2022-11-01 21:18:46 caching file sra.sra.SRP009615.MD.gz.
#> 2022-11-01 21:18:46 caching file sra.recount_project.SRP009615.MD.gz.
#> 2022-11-01 21:18:46 caching file sra.recount_qc.SRP009615.MD.gz.
#> 2022-11-01 21:18:46 caching file sra.recount_seq_qc.SRP009615.MD.gz.
#> 2022-11-01 21:18:46 caching file sra.recount_pred.SRP009615.MD.gz.
#> 2022-11-01 21:18:46 downloading and reading the feature information.
#> 2022-11-01 21:18:46 caching file human.gene_sums.G026.gtf.gz.
#> 2022-11-01 21:18:47 downloading and reading the counts: 12 samples across 63856 features.
#> 2022-11-01 21:18:47 caching file sra.gene_sums.SRP009615.G026.gz.
#> 2022-11-01 21:18:47 construcing the RangedSummarizedExperiment (rse) object.

## Expand the SRA attributes (see details for more information)
rse_gene_SRP009615 <- expand_sra_attributes(rse_gene_SRP009615)