Find available sub-directories for a given URL

This function lists all the available files inside a given URL. If url is a local directory, such as when you download all the recount3 files to your computer, then this function will list all the files inside url.

file_list(url, bfc = BiocFileCache::BiocFileCache())

Arguments

url
A character(1) with either a URL or a local file path.

bfc
A BiocFileCache-class object where the files will be cached to.

Value

A character() vector with the available options.

Examples

## List the contents in a URL
file_list(
    "http://snaptron.cs.jhu.edu/data/temp/recount3/human/data_sources/"
)

#> adding rname 'http://snaptron.cs.jhu.edu/data/temp/recount3/human/data_sources/'

#> [1] "gtex" "sra" 

## List the contents in a directory
file_list(getwd())

#>  [1] "annotation_ext.html"             "annotation_options.html"        
#>  [3] "available_projects.html"         "available_samples.html"         
#>  [5] "compute_read_counts-1.png"       "compute_read_counts-2.png"      
#>  [7] "compute_read_counts.html"        "compute_scale_factors-1.png"    
#>  [9] "compute_scale_factors.html"      "create_rse_manual.html"         
#> [11] "create_rse.html"                 "figures"                        
#> [13] "file_list.html"                  "file_locate_url_annotation.html"
#> [15] "file_locate_url.html"            "file_retrieve.html"             
#> [17] "foo.html"                        "index.html"                     
#> [19] "is_paired_end.html"              "project_home_available.html"    
#> [21] "read_counts.html"                "read_metadata.html"             
#> [23] "recount3_hub.html"               "recount3-package.html"          
#> [25] "Rplot001.png"                    "Rplot002.png"                   
#> [27] "transform_counts.html"