Download a remote file and cache it to re-use later

Download a remote file and cache it to re-use later

file_retrieve(url, bfc = BiocFileCache::BiocFileCache())

Arguments

url
A character(1) with the file URL or the actual local path in which case, it won't be cached. If length(url) > 1 , this function will be used recursively.

bfc
A BiocFileCache-class object where the files will be cached to.

Value

A character(1) with the path to the cached file.

See also

Examples

## Download the metadata file for project SRP009615
url_SRP009615_meta <- file_locate_url(
    "SRP009615",
    "data_sources/sra"
)
local_SRP009615_meta <- file_retrieve(
    url = url_SRP009615_meta
)

#> 2020-10-01 15:13:43 caching file sra.sra.SRP009615.MD.gz.

#> 2020-10-01 15:13:44 caching file sra.recount_project.SRP009615.MD.gz.

#> 2020-10-01 15:13:44 caching file sra.recount_qc.SRP009615.MD.gz.

#> 2020-10-01 15:13:44 caching file sra.recount_seq_qc.SRP009615.MD.gz.

#> 2020-10-01 15:13:44 caching file sra.recount_pred.SRP009615.MD.gz.

local_SRP009615_meta

#>                                                          sra.sra.SRP009615.MD.gz 
#>             "/tmp/RtmpyQ4ziy/BiocFileCache/24da17985f0c_sra.sra.SRP009615.MD.gz" 
#>                                              sra.recount_project.SRP009615.MD.gz 
#> "/tmp/RtmpyQ4ziy/BiocFileCache/24da6bf27903_sra.recount_project.SRP009615.MD.gz" 
#>                                                   sra.recount_qc.SRP009615.MD.gz 
#>       "/tmp/RtmpyQ4ziy/BiocFileCache/24dafb3feb3_sra.recount_qc.SRP009615.MD.gz" 
#>                                               sra.recount_seq_qc.SRP009615.MD.gz 
#>   "/tmp/RtmpyQ4ziy/BiocFileCache/24dab6e059b_sra.recount_seq_qc.SRP009615.MD.gz" 
#>                                                 sra.recount_pred.SRP009615.MD.gz 
#>    "/tmp/RtmpyQ4ziy/BiocFileCache/24da2f1a758a_sra.recount_pred.SRP009615.MD.gz" 

## Download the gene counts file for project SRP009615
url_SRP009615_gene <- file_locate_url(
    "SRP009615",
    "data_sources/sra",
    type = "gene"
)
local_SRP009615_gene <- file_retrieve(
    url = url_SRP009615_gene
)

#> 2020-10-01 15:13:44 caching file sra.gene_sums.SRP009615.G026.gz.

local_SRP009615_gene

#>                                              sra.gene_sums.SRP009615.G026.gz 
#> "/tmp/RtmpyQ4ziy/BiocFileCache/24da63067fe0_sra.gene_sums.SRP009615.G026.gz"