Construct the URL to a recount3 annotation file

Given a expression feature type, organism and annotation, this function constructs the URL (or file path) to access a recount3 annotation file. This function is used by create_rse_manual().

locate_url_ann(
  type = c("gene", "exon"),
  organism = c("human", "mouse"),
  annotation = annotation_options(organism),
  recount3_url = getOption("recount3_url", "http://duffel.rail.bio/recount3")
)

Arguments

type

A character(1) specifying whether you want to access gene counts or exon data.

organism

A character(1) specifying which organism you want to download data from. Supported options are "human" or "mouse".

annotation

A character(1) specifying which annotation you want to download. Only used when type is either gene or exon.

recount3_url

A character(1) specifying the home URL for recount3 or a local directory where you have mirrored recount3.

Value

A character(1) with the URL (or file path) to access the recount3 annotation file.

See also

Examples

locate_url_ann()
#>                                                               human.gene_sums.G026.gtf.gz 
#> "http://duffel.rail.bio/recount3/human/annotations/gene_sums/human.gene_sums.G026.gtf.gz" 
locate_url_ann(organism = "mouse")
#>                                                               mouse.gene_sums.M023.gtf.gz 
#> "http://duffel.rail.bio/recount3/mouse/annotations/gene_sums/mouse.gene_sums.M023.gtf.gz"