Find available project home options

This function finds the home for a given project (study) of interest based on the organism and the home_type.

project_homes(
  organism = c("human", "mouse"),
  recount3_url = getOption("recount3_url", "http://duffel.rail.bio/recount3")
)

Arguments

organism

A character(1) specifying which organism you want to download data from. Supported options are "human" or "mouse".

recount3_url

A character(1) specifying the home URL for recount3 or a local directory where you have mirrored recount3. Defaults to the load balancer http://duffel.rail.bio/recount3, but can also be https://recount-opendata.s3.amazonaws.com/recount3/release from https://registry.opendata.aws/recount/ or SciServer datascope from IDIES at JHU https://sciserver.org/public-data/recount3/data. You can set the R option recount3_url (for example in your .Rprofile) if you have a favorite mirror.

Value

A character() vector with the available project_home options.

Details

By default it reads a small text file from recount3_url/organism/homes_index using readLines(). This text file should contain each possible project home per line. See http://duffel.rail.bio/recount3/human/homes_index for an example.

See also

Examples

## List the different available `project_home` options for the default
## arguments
project_homes("human")
#> [1] "data_sources/sra"  "data_sources/gtex" "data_sources/tcga"
project_homes("mouse")
#> [1] "data_sources/sra"

## Test files
project_homes("human",
    recount3_url = "http://snaptron.cs.jhu.edu/data/temp/recount3"
)
#> [1] "data_sources/sra"              "data_sources/gtex"            
#> [3] "collections/ccle"              "collections/geuvadis_smartseq"
#> [5] "collections/gtex_geuvadis"