Read a metadata file

This function reads in a recount3 metadata file into R. You can first locate the file using file_locate_url() then download it to your computer using file_retrieve().

read_metadata(metadata_file)

Arguments

metadata_file
A character(1) with the local path to a recount3 metadata file.

Value

A data.frame() with all lower case column names for the sample metadata.

Examples

## Download the metadata file for project ERP001942
url_ERP001942_meta <- file_locate_url("ERP001942", "data_sources/sra")
local_ERP001942_meta <- file_retrieve(url = url_ERP001942_meta)

## Read the metadata
ERP001942_meta <- read_metadata(local_ERP001942_meta)

## Locate and read the GTEx bladder metadata
gtex_bladder_meta <- read_metadata(
    file_retrieve(
        file_locate_url("BLADDER", "data_sources/gtex")
    )
)

#> adding rname 'http://snaptron.cs.jhu.edu/data/temp/recount3/human/data_sources/gtex/metadata/ER/BLADDER/gtex.metadata.BLADDER.MD.gz'

head(gtex_bladder_meta)

#>   rail_id    run_acc              rail_barcode    subjid sex   age dthhrdy
#> 1  153881 SRR1071717 GTEX-U3ZM-0826-SM-4DXU6.1 GTEX-U3ZM   1 40-49       0
#> 2  309816 SRR1079830 GTEX-U4B1-1226-SM-4DXT7.1 GTEX-U4B1   1 40-49       0
#> 3  270898 SRR1081765 GTEX-S4Q7-0926-SM-4AD5D.1 GTEX-S4Q7   1 20-29       0
#> 4  615530 SRR1085402 GTEX-SNMC-0826-SM-4DM66.1 GTEX-SNMC   1 20-29       0
#> 5  607337 SRR1086236 GTEX-TMMY-1526-SM-4DXST.1 GTEX-TMMY   2 40-49       0
#> 6  277053 SRR1092208 GTEX-SE5C-1026-SM-4BRUG.1 GTEX-SE5C   2 40-49       0
#>                    sampid smatsscr smcenter
#> 1 GTEX-U3ZM-0826-SM-4DXU6        2       B1
#> 2 GTEX-U4B1-1226-SM-4DXT7        1       B1
#> 3 GTEX-S4Q7-0926-SM-4AD5D        1       B1
#> 4 GTEX-SNMC-0826-SM-4DM66        2       B1
#> 5 GTEX-TMMY-1526-SM-4DXST        2       C1
#> 6 GTEX-SE5C-1026-SM-4BRUG        1       B1
#>                                                                                                   smpthnts
#> 1 6 pieces  ~10x4mm.  Urothelium variably sloughed, better preserved than usual, less than 1% of thickness
#> 2                                                         8 pieces, up to 10x3 mm; preserved mucosa on all
#> 3                          7 pieces,  ~7x3.5mm.  Urothelium remarkably well-preserved,  ~2-3% of thickness
#> 4                   6 pieces up to 8x4 mm; mucosa is variably preserved: some areas well and some sloughed
#> 5                            6 pieces  ~7x5mm.  Urothelium partly sloughed but better preserved than usual
#> 6                                                         6 pieces up to 8x5 mm; only one piece has mucosa
#>   smrin    smts   smtsd smubrid smtsisch smtspax smnabtch
#> 1   7.5 Bladder Bladder    1255      230     747 BP-25560
#> 2   7.8 Bladder Bladder    1255      191    1225 BP-25601
#> 3   7.7 Bladder Bladder    1255      161     412 BP-23051
#> 4   8.6 Bladder Bladder    1255      230    1045 BP-23794
#> 5   8.0 Bladder Bladder    1255      439    1181 BP-25190
#> 6   7.7 Bladder Bladder    1255      185     972 BP-23389
#>                            smnabtcht  smnabtchd   smgebtch  smgebtchd smgebtcht
#> 1 RNA isolation_PAXgene Tissue miRNA 05/22/2012 LCSET-3206 06/04/2013 TruSeq.v1
#> 2 RNA isolation_PAXgene Tissue miRNA 05/23/2012 LCSET-3206 06/04/2013 TruSeq.v1
#> 3 RNA isolation_PAXgene Tissue miRNA 03/13/2012 LCSET-3098 05/14/2013 TruSeq.v1
#> 4 RNA isolation_PAXgene Tissue miRNA 04/04/2012 LCSET-3205 05/31/2013 TruSeq.v1
#> 5 RNA isolation_PAXgene Tissue miRNA 05/09/2012 LCSET-3206 06/04/2013 TruSeq.v1
#> 6 RNA isolation_PAXgene Tissue miRNA 03/26/2012 LCSET-3152 05/27/2013 TruSeq.v1
#>   smafrze smgtc sme2mprt smchmprs smntrart smnumgps  smmaprt smexncrt sm550nrm
#> 1  RNASEQ    NA 0.988668   339167 0.975561       NA 0.991069 0.843516       NA
#> 2  RNASEQ    NA 0.988598   536713 0.973821       NA 0.991280 0.821863       NA
#> 3  RNASEQ    NA 0.990838   290714 0.973011       NA 0.993009 0.801211       NA
#> 4  RNASEQ    NA 0.988295   293534 0.967582       NA 0.990607 0.763716       NA
#> 5  RNASEQ    NA 0.988508   369265 0.975620       NA 0.991073 0.823269       NA
#> 6  RNASEQ    NA 0.991046   394834 0.976320       NA 0.993359 0.848322       NA
#>   smgnsdtc smunmprt sm350nrm smrdlgth smmncpb   sme1mmrt smsflgth smestlbs
#> 1    21262        1       NA       76      NA 0.00209364      139        0
#> 2    22592        1       NA       76      NA 0.00224570      157        0
#> 3    23205        1       NA       76      NA 0.00202882      131        0
#> 4    23092        1       NA       76      NA 0.00208335      110        0
#> 5    22184        1       NA       76      NA 0.00222215      144        0
#> 6    21412        1       NA       76      NA 0.00221069      158        0
#>     smmppd  smnterrt smrrnanm  smrdttl smvqcfl smmncv smtrscpt smmppdpr
#> 1 63926700 0.0241488   738605 64502700 5839470     NA    21277 31792900
#> 2 73718500 0.0258600   374650 74367000 6567550     NA    22622 36646200
#> 3 61337000 0.0266825   212377 61768800 3683040     NA    23225 30529500
#> 4 51773500 0.0321327   235558 52264400 3637220     NA    23118 25760400
#> 5 68070100 0.0240411   283345 68683200 6200030     NA    22201 33853700
#> 6 72052900 0.0234005   491168 72534500 7105490     NA    21427 35856700
#>   smcglgth smgappct smunpdrd smntrnrt smmpunrt smexpeff smmppdun   sme2mmrt
#> 1       NA       NA        0 0.132045 0.991069 0.835983 63926700 0.00297382
#> 2       NA       NA        0 0.151958 0.991280 0.814696 73718500 0.00313601
#> 3       NA       NA        0 0.171800 0.993009 0.795610 61337000 0.00277657
#> 4       NA       NA        0 0.203866 0.990607 0.756543 51773500 0.00345301
#> 5       NA       NA        0 0.152351 0.991073 0.815920 68070100 0.00305874
#> 6       NA       NA        0 0.127998 0.993359 0.842688 72052900 0.00291208
#>   sme2anti smaltalg sme2snse smmflgth sme1anti smspltrd   smbsmmrt sme1snse
#> 1 13807000  5680860 13836800      159 13810400 11069400 0.00253266 13979700
#> 2 15965200  6506070 15978200      169 16007900 13524600 0.00268965 16124700
#> 3 13458000  5397620 13489700      141 13510700 10989200 0.00240188 13565700
#> 4 11429000  6115360 11465000      112 11507200  9054400 0.00276658 11502900
#> 5 14772000  6248800 14774700      154 14798900 12928200 0.00263936 14915200
#> 6 15389200  5930350 15519600      172 15501800 13665500 0.00256057 15558500
#>   sme1pcts   smrrnart sme1mprt smnum5cd smdpmprt sme2pcts junction_count
#> 1  50.3047 0.01145080 0.993470       NA        0  50.0538         247600
#> 2  50.1818 0.00503785 0.993962       NA        0  50.0203         188783
#> 3  50.1015 0.00343826 0.995181       NA        0  50.0589         302563
#> 4  49.9907 0.00450704 0.992919       NA        0  50.0785         253719
#> 5  50.1956 0.00412539 0.993638       NA        0  50.0046         227259
#> 6  50.0913 0.00677150 0.995672       NA        0  50.2108         458369
#>   junction_coverage junction_avg_coverage
#> 1           7282179              29.41106
#> 2          15845256              83.93370
#> 3          16243715              53.68705
#> 4          13421701              52.89987
#> 5          16123862              70.94928
#> 6          10194323              22.24043