Access recount3 project_info objects through ExperimentHub

This function downloads from ExperimentHub the list of available recount3 projects. However, you might be more tempted to explore the data interactively using code like eh2 <- AnnotationHub::display(ExperimentHub::ExperimentHub())

recount3_hub(
  project,
  data_source = c("sra", "gtex", "tcga"),
  organism = c("human", "mouse"),
  eh = ExperimentHub::ExperimentHub()
)

Arguments

project
A character(1) specifying the project ID you are interested in. Could be a SRA study ID or could be a specific GTEx or TCGA tissue.

data_source
A character(1) specifying the data source.

organism
A character(1) specifying which organism you want to download data from. Supported options are "human" or "mouse".

eh
An ExperimentHub object ExperimentHub-class.

Value

A data.frame() with one row and three columns that contain the information needed by create_rse() for accessing the data.

Examples

## Locate a project of interest
project_info <- recount3_hub("SRP009615")

#> snapshotDate(): 2020-09-24

#> 2020-09-28 04:49:40 the project SRP009615 was not found on ExperimentHub.
#> Using a backup mechanism powered by recount3::available_projects().

#> 2020-09-28 04:49:40 caching file sra.recount_project.MD.gz.

#> 2020-09-28 04:49:40 caching file gtex.recount_project.MD.gz.

#> 2020-09-28 04:49:40 caching file tcga.recount_project.MD.gz.

## For an interactive display use:
if (interactive()) {
    ## Open an interactive display
    eh2 <- AnnotationHub::display(ExperimentHub::ExperimentHub())

    ## Select one study, then hit the "send" button on the interactive
    ## display

    ## Check that you selected one study
    if (length(eh2) == 1) {
        ## Download the data using the ExperimentHub [[ syntax
        project_info <- eh2[[1]]
    }
}

## Create a RangedSummarizedExperiment object for this project
create_rse(project_info)

#> 2020-09-28 04:49:45 downloading and reading the metadata.

#> 2020-09-28 04:49:45 caching file sra.sra.SRP009615.MD.gz.

#> 2020-09-28 04:49:45 caching file sra.recount_project.SRP009615.MD.gz.

#> 2020-09-28 04:49:45 caching file sra.recount_qc.SRP009615.MD.gz.

#> 2020-09-28 04:49:45 caching file sra.recount_seq_qc.SRP009615.MD.gz.

#> 2020-09-28 04:49:46 caching file sra.recount_pred.SRP009615.MD.gz.

#> Warning: Replacing study_acc by study in sra.sra.SRP009615.MD.gz

#> Warning: Replacing run_acc by external_id in sra.sra.SRP009615.MD.gz

#> Warning: Replacing study_acc by study in sra.recount_qc.SRP009615.MD.gz

#> Warning: Replacing run_acc by external_id in sra.recount_qc.SRP009615.MD.gz

#> 2020-09-28 04:49:46 downloading and reading the feature information.

#> 2020-09-28 04:49:46 caching file human.gene_sums.G026.gtf.gz.

#> 2020-09-28 04:49:46 downloading and reading the counts: 12 samples across 63856 features.

#> 2020-09-28 04:49:47 caching file sra.gene_sums.SRP009615.G026.gz.

#> 2020-09-28 04:49:47 construcing the RangedSummarizedExperiment (rse) object.

#> class: RangedSummarizedExperiment 
#> dim: 63856 12 
#> metadata(8): time_created recount3_version ... annotation recount3_url
#> assays(1): raw_counts
#> rownames(63856): ENSG00000278704.1 ENSG00000277400.1 ...
#>   ENSG00000182484.15_PAR_Y ENSG00000227159.8_PAR_Y
#> rowData names(10): source type ... havana_gene tag
#> colnames(12): SRR387777 SRR387778 ... SRR389077 SRR389078
#> colData names(173): rail_id external_id ... recount_pred.pred.type
#>   BigWigURL