List available projects in recount3

List available projects in recount3

available_projects(
  organism = c("human", "mouse"),
  recount3_url = "http://idies.jhu.edu/recount3/data",
  bfc = BiocFileCache::BiocFileCache()
)

Arguments

organism
A character(1) specifying which organism you want to download data from. Supported options are "human" or "mouse".

recount3_url
A character(1) specifying the home URL for recount3 or a local directory where you have mirrored recount3.

bfc
A BiocFileCache-class object where the files will be cached to.

Value

A data.frame() with the project ID (project), the organism, the file_source from where the data was accessed, the recount3 project home location (project_home), and the project project_type that differentiates between data_sources and compilations.

Examples

## Find all the human projects
human_projects <- available_projects()

#> using temporary cache /tmp/RtmpZJUtoN/BiocFileCache

#> 2020-10-02 12:28:28 caching file sra.recount_project.MD.gz.

#> adding rname 'http://idies.jhu.edu/recount3/data/human/data_sources/sra/metadata/sra.recount_project.MD.gz'

#> 2020-10-02 12:28:29 caching file gtex.recount_project.MD.gz.

#> adding rname 'http://idies.jhu.edu/recount3/data/human/data_sources/gtex/metadata/gtex.recount_project.MD.gz'

#> 2020-10-02 12:28:29 caching file tcga.recount_project.MD.gz.

#> adding rname 'http://idies.jhu.edu/recount3/data/human/data_sources/tcga/metadata/tcga.recount_project.MD.gz'

## Explore the results
dim(human_projects)

#> [1] 8742    5

head(human_projects)

#>       project organism file_source     project_home project_type
#> 1   SRP107565    human         sra data_sources/sra data_sources
#> 217 SRP149665    human         sra data_sources/sra data_sources
#> 221 SRP017465    human         sra data_sources/sra data_sources
#> 244 SRP119165    human         sra data_sources/sra data_sources
#> 250 SRP133965    human         sra data_sources/sra data_sources
#> 262 SRP096765    human         sra data_sources/sra data_sources

## How many are from a data source vs a compilation?
table(human_projects$project_type, useNA = "ifany")

#> 
#> data_sources 
#>         8742 

## What are the unique file sources?
table(
    human_projects$file_source[human_projects$project_type == "data_sources"]
)

#> 
#> gtex  sra tcga 
#>   32 8677   33 

## Note that big projects are broken up to make them easier to access
## For example, GTEx and TCGA are broken up by tissue
head(subset(human_projects, file_source == "gtex"))

#>               project organism file_source      project_home project_type
#> 316450 ADIPOSE_TISSUE    human        gtex data_sources/gtex data_sources
#> 316451         MUSCLE    human        gtex data_sources/gtex data_sources
#> 316452   BLOOD_VESSEL    human        gtex data_sources/gtex data_sources
#> 316454          HEART    human        gtex data_sources/gtex data_sources
#> 316456          OVARY    human        gtex data_sources/gtex data_sources
#> 316457         UTERUS    human        gtex data_sources/gtex data_sources

head(subset(human_projects, file_source == "tcga"))

#>        project organism file_source      project_home project_type
#> 335664     ACC    human        tcga data_sources/tcga data_sources
#> 335743    BLCA    human        tcga data_sources/tcga data_sources
#> 336176    BRCA    human        tcga data_sources/tcga data_sources
#> 337432    CESC    human        tcga data_sources/tcga data_sources
#> 337741    CHOL    human        tcga data_sources/tcga data_sources
#> 337786    COAD    human        tcga data_sources/tcga data_sources

## Find all the mouse projects
mouse_projects <- available_projects(organism = "mouse")

#> 2020-10-02 12:28:35 caching file sra.recount_project.MD.gz.

#> adding rname 'http://idies.jhu.edu/recount3/data/mouse/data_sources/sra/metadata/sra.recount_project.MD.gz'

## Explore the results
dim(mouse_projects)

#> [1] 10088     5

head(mouse_projects)

#>       project organism file_source     project_home project_type
#> 1   SRP170963    mouse         sra data_sources/sra data_sources
#> 370 SRP045763    mouse         sra data_sources/sra data_sources
#> 392 DRP005463    mouse         sra data_sources/sra data_sources
#> 404 SRP172863    mouse         sra data_sources/sra data_sources
#> 412 SRP186363    mouse         sra data_sources/sra data_sources
#> 425 SRP101363    mouse         sra data_sources/sra data_sources

## How many are from a data source vs a compilation?
table(mouse_projects$project_type, useNA = "ifany")

#> 
#> data_sources 
#>        10088 

## What are the unique file sources?
table(
    mouse_projects$file_source[mouse_projects$project_type == "data_sources"]
)

#> 
#>   sra 
#> 10088