List available samples in recount3

This function returns a data.frame() with the samples that are available from recount3. Note that a specific sample might be available from a given data_source and none or many collections.

available_samples(
  organism = c("human", "mouse"),
  recount3_url = "http://idies.jhu.edu/recount3/data",
  bfc = BiocFileCache::BiocFileCache()
)

Arguments

organism
A character(1) specifying which organism you want to download data from. Supported options are "human" or "mouse".

recount3_url
A character(1) specifying the home URL for recount3 or a local directory where you have mirrored recount3.

bfc
A BiocFileCache-class object where the files will be cached to.

Value

A data.frame() with the sample ID used by the original source of the data (external_id), the project ID (project), the organism, the file_source from where the data was accessed, the date the sample was processed (date_processed) in YYYY-MM-DD format, the recount3 project home location (project_home), and the project project_type that differentiates between data_sources and compilations.

Examples

## Find all the human samples available from recount3
human_samples <- available_samples()

#> 2020-10-02 21:09:05 caching file sra.recount_project.MD.gz.

#> 2020-10-02 21:09:06 caching file gtex.recount_project.MD.gz.

#> 2020-10-02 21:09:06 caching file tcga.recount_project.MD.gz.

dim(human_samples)

#> [1] 347011      7

head(human_samples)

#>   external_id   project organism file_source date_processed     project_home
#> 1  SRR5579327 SRP107565    human         sra     2019-10-01 data_sources/sra
#> 2  SRR5579328 SRP107565    human         sra     2019-10-01 data_sources/sra
#> 3  SRR5579329 SRP107565    human         sra     2019-10-01 data_sources/sra
#> 4  SRR5579330 SRP107565    human         sra     2019-10-01 data_sources/sra
#> 5  SRR5579331 SRP107565    human         sra     2019-10-01 data_sources/sra
#> 6  SRR5579332 SRP107565    human         sra     2019-10-01 data_sources/sra
#>   project_type
#> 1 data_sources
#> 2 data_sources
#> 3 data_sources
#> 4 data_sources
#> 5 data_sources
#> 6 data_sources

## How many are from a data source vs a compilation?
table(human_samples$project_type, useNA = "ifany")

#> 
#> data_sources 
#>       347011 

## What are the unique file sources?
table(
    human_samples$file_source[human_samples$project_type == "data_sources"]
)

#> 
#>   gtex    sra   tcga 
#>  19214 316449  11348 

## Find all the mouse samples available from recount3
mouse_samples <- available_samples("mouse")

#> 2020-10-02 21:09:10 caching file sra.recount_project.MD.gz.

dim(mouse_samples)

#> [1] 416903      7

head(mouse_samples)

#>   external_id   project organism file_source date_processed     project_home
#> 1  SRR8249198 SRP170963    mouse         sra     2020-01-01 data_sources/sra
#> 2  SRR8249199 SRP170963    mouse         sra     2020-01-01 data_sources/sra
#> 3  SRR8249200 SRP170963    mouse         sra     2020-01-01 data_sources/sra
#> 4  SRR8249201 SRP170963    mouse         sra     2020-01-01 data_sources/sra
#> 5  SRR8249202 SRP170963    mouse         sra     2020-01-01 data_sources/sra
#> 6  SRR8249205 SRP170963    mouse         sra     2020-01-01 data_sources/sra
#>   project_type
#> 1 data_sources
#> 2 data_sources
#> 3 data_sources
#> 4 data_sources
#> 5 data_sources
#> 6 data_sources

## How many are from a data source vs a compilation?
table(mouse_samples$project_type, useNA = "ifany")

#> 
#> data_sources 
#>       416903