Create a recount3 RangedSummarizedExperiment gene or exon object

Once you have identified a project you want to work with, you can use this function to construct a recount3 RangedSummarizedExperiment-class (RSE) object at the gene or exon expression feature level. This function will retrieve the data, cache it, then assemble the RSE object.

create_rse(
  project_info,
  type = c("gene", "exon"),
  annotation = annotation_options(project_info$organism),
  bfc = BiocFileCache::BiocFileCache(),
  recount3_url = "http://snaptron.cs.jhu.edu/data/temp/recount3"
)

Arguments

project_info
A data.frame() with one row that contains the information for the project you are interested in. You can find which project to work on using todo().

type
A character(1) specifying whether you want to access gene counts or exon counts.

annotation
A character(1) specifying which annotation you want to download. Only used when type is either gene or exon.

bfc
A BiocFileCache-class object where the files will be cached to.

recount3_url
A character(1) specifying the home URL for recount3 or a local directory where you have mirrored recount3.

Value

A RangedSummarizedExperiment-class object.

Examples

## Find the project you are interested in
## (temporarily build this data.frame)
proj_info <- data.frame(
    project = "ERP110066",
    project_home = "data_sources/sra",
    organism = "human"
)

## Create a RSE object at the gene level
rse_gene_ERP110066 <- create_rse(proj_info)

#> using temporary cache /tmp/RtmpXX0Qcp/BiocFileCache

#> adding rname 'http://snaptron.cs.jhu.edu/data/temp/recount3/human/data_sources'

#> adding rname 'http://snaptron.cs.jhu.edu/data/temp/recount3/human/collections'

#> 2020-05-27 05:24:02 downloading and reading the metadata

#> Caching file sra.ncbi.ERP110066.MD.gz

#> adding rname 'http://snaptron.cs.jhu.edu/data/temp/recount3/human/data_sources/sra/metadata/66/ERP110066/sra.ncbi.ERP110066.MD.gz'

#> Caching file sra.recount_project.ERP110066.MD.gz

#> adding rname 'http://snaptron.cs.jhu.edu/data/temp/recount3/human/data_sources/sra/metadata/66/ERP110066/sra.recount_project.ERP110066.MD.gz'

#> Caching file sra.recount_qc.ERP110066.MD.gz

#> adding rname 'http://snaptron.cs.jhu.edu/data/temp/recount3/human/data_sources/sra/metadata/66/ERP110066/sra.recount_qc.ERP110066.MD.gz'

#> Warning: download failed
#>   web resource path: ‘http://snaptron.cs.jhu.edu/data/temp/recount3/human/data_sources/sra/metadata/66/ERP110066/sra.recount_qc.ERP110066.MD.gz’
#>   local file path: ‘/tmp/RtmpXX0Qcp/BiocFileCache/be11e7f62e5_sra.recount_qc.ERP110066.MD.gz’
#>   reason: Timeout was reached: [snaptron.cs.jhu.edu] Connection timed out after 10001 milliseconds

#> Warning: bfcadd() failed; resource removed
#>   rid: BFC5
#>   fpath: ‘http://snaptron.cs.jhu.edu/data/temp/recount3/human/data_sources/sra/metadata/66/ERP110066/sra.recount_qc.ERP110066.MD.gz’
#>   reason: download failed

#> Warning: 
#> trying to add rname 'http://snaptron.cs.jhu.edu/data/temp/recount3/human/data_sources/sra/metadata/66/ERP110066/sra.recount_qc.ERP110066.MD.gz' produced error:
#>   bfcadd() failed; see warnings()

#> Error in BiocFileCache::bfcrpath(bfc, url, exact = TRUE): not all 'rnames' found or unique.

## Explore the resulting RSE gene object
rse_gene_ERP110066

#> Error in eval(expr, envir, enclos): object 'rse_gene_ERP110066' not found

dim(rse_gene_ERP110066)

#> Error in eval(expr, envir, enclos): object 'rse_gene_ERP110066' not found

rowRanges(rse_gene_ERP110066)

#> Error in rowRanges(rse_gene_ERP110066): object 'rse_gene_ERP110066' not found

colnames(colData(rse_gene_ERP110066))

#> Error in colData(rse_gene_ERP110066): object 'rse_gene_ERP110066' not found

## Create a second RSE object using another annotation
rse_gene_ERP110066_gencode_v29 <- create_rse(
    proj_info,
    annotation = "gencode_v29"
)

#> 2020-05-27 05:24:46 downloading and reading the metadata

#> Caching file sra.ncbi.ERP110066.MD.gz

#> Caching file sra.recount_project.ERP110066.MD.gz

#> Caching file sra.recount_qc.ERP110066.MD.gz

#> adding rname 'http://snaptron.cs.jhu.edu/data/temp/recount3/human/data_sources/sra/metadata/66/ERP110066/sra.recount_qc.ERP110066.MD.gz'

#> 2020-05-27 05:24:48 downloading and reading the feature information

#> Caching file human.gene_sums.G029.gtf.gz

#> adding rname 'http://snaptron.cs.jhu.edu/data/temp/recount3/human/annotations/gene_sums/human.gene_sums.G029.gtf.gz'

#> 2020-05-27 05:24:49 downloading and reading the counts: 1024 samples across 64837 features.

#> Caching file sra.gene_sums.ERP110066.G029.gz

#> adding rname 'http://snaptron.cs.jhu.edu/data/temp/recount3/human/data_sources/sra/gene_sums/66/ERP110066/sra.gene_sums.ERP110066.G029.gz'

#> 2020-05-27 05:24:57 construcing the RangedSummarizedExperiment (rse) object

## Doesn't work right now
# rse_gene_ERP110066_refseq <- create_rse(
#     proj_info,
#     annotation = "refseq"
# )

## Create a RSE object at the exon level
if (FALSE) {
rse_exon_ERP110066 <- create_rse(
    proj_info,
    type = "exon"
)

## Explore the resulting RSE exon object
rse_exon_ERP110066

dim(rse_exon_ERP110066)
rowRanges(rse_exon_ERP110066)
colData(rse_exon_ERP110066)
colnames(colData(rse_exon_ERP110066))
}