Create a recount3 RangedSummarizedExperiment gene or exon object

Once you have identified a project you want to work with, you can use this function to construct a recount3 RangedSummarizedExperiment-class (RSE) object at the gene or exon expression feature level. This function will retrieve the data, cache it, then assemble the RSE object.

create_rse(
  project_info,
  type = c("gene", "exon"),
  annotation = annotation_options(project_info$organism),
  bfc = BiocFileCache::BiocFileCache(),
  recount3_url = "http://snaptron.cs.jhu.edu/data/temp/recount3"
)

Arguments

project_info
A data.frame() with one row that contains the information for the project you are interested in. You can find which project to work on using todo().

type
A character(1) specifying whether you want to access gene counts or exon counts.

annotation
A character(1) specifying which annotation you want to download. Only used when type is either gene or exon.

bfc
A BiocFileCache-class object where the files will be cached to.

recount3_url
A character(1) specifying the home URL for recount3 or a local directory where you have mirrored recount3.

Value

A RangedSummarizedExperiment-class object.

Examples

## Find the project you are interested in
## (temporarily build this data.frame)
proj_info <- data.frame(
    project = "ERP110066",
    project_home = "data_sources/sra",
    organism = "human"
)

## Create a RSE object at the gene level
rse_gene_ERP110066 <- create_rse(proj_info)

#> using temporary cache /tmp/RtmpvzkdzB/BiocFileCache

#> adding rname 'http://snaptron.cs.jhu.edu/data/temp/recount3/human/data_sources'

#> adding rname 'http://snaptron.cs.jhu.edu/data/temp/recount3/human/collections'

#> 2020-05-27 01:36:44 downloading and reading the metadata

#> Caching file sra.ncbi.ERP110066.MD.gz

#> adding rname 'http://snaptron.cs.jhu.edu/data/temp/recount3/human/data_sources/sra/metadata/66/ERP110066/sra.ncbi.ERP110066.MD.gz'

#> Caching file sra.recount_project.ERP110066.MD.gz

#> adding rname 'http://snaptron.cs.jhu.edu/data/temp/recount3/human/data_sources/sra/metadata/66/ERP110066/sra.recount_project.ERP110066.MD.gz'

#> Caching file sra.recount_qc.ERP110066.MD.gz

#> adding rname 'http://snaptron.cs.jhu.edu/data/temp/recount3/human/data_sources/sra/metadata/66/ERP110066/sra.recount_qc.ERP110066.MD.gz'

#> 2020-05-27 01:36:47 downloading and reading the feature information

#> Caching file human.gene_sums.G026.gtf.gz

#> adding rname 'http://snaptron.cs.jhu.edu/data/temp/recount3/human/annotations/gene_sums/human.gene_sums.G026.gtf.gz'

#> 2020-05-27 01:36:48 downloading and reading the counts: 1024 samples across 63856 features.

#> Caching file sra.gene_sums.ERP110066.G026.gz

#> adding rname 'http://snaptron.cs.jhu.edu/data/temp/recount3/human/data_sources/sra/gene_sums/66/ERP110066/sra.gene_sums.ERP110066.G026.gz'

#> 2020-05-27 01:36:56 construcing the RangedSummarizedExperiment (rse) object

## Explore the resulting RSE gene object
rse_gene_ERP110066

#> class: RangedSummarizedExperiment 
#> dim: 63856 1024 
#> metadata(0):
#> assays(1): counts
#> rownames(63856): ENSG00000278704.1 ENSG00000277400.1 ...
#>   ENSG00000182484.15_PAR_Y ENSG00000227159.8_PAR_Y
#> rowData names(10): source type ... havana_gene tag
#> colnames(1024): ERR2713604 ERR2713605 ... ERR2713998 ERR2713999
#> colData names(139): rail_id run_acc ...
#>   recount_qc.junction_avg_coverage BigWigURL

dim(rse_gene_ERP110066)

#> [1] 63856  1024

rowRanges(rse_gene_ERP110066)

#> GRanges object with 63856 ranges and 10 metadata columns:
#>                              seqnames            ranges strand |   source
#>                                 <Rle>         <IRanges>  <Rle> | <factor>
#>          ENSG00000278704.1 GL000009.2       56140-58376      - |  ENSEMBL
#>          ENSG00000277400.1 GL000194.1      53590-115018      - |  ENSEMBL
#>          ENSG00000274847.1 GL000194.1      53594-115055      - |  ENSEMBL
#>          ENSG00000277428.1 GL000195.1       37434-37534      - |  ENSEMBL
#>          ENSG00000276256.1 GL000195.1       42939-49164      - |  ENSEMBL
#>                        ...        ...               ...    ... .      ...
#>   ENSG00000124334.17_PAR_Y       chrY 57184101-57197337      + |   HAVANA
#>   ENSG00000185203.12_PAR_Y       chrY 57201143-57203357      - |   HAVANA
#>    ENSG00000270726.6_PAR_Y       chrY 57190738-57208756      + |   HAVANA
#>   ENSG00000182484.15_PAR_Y       chrY 57207346-57212230      + |   HAVANA
#>    ENSG00000227159.8_PAR_Y       chrY 57212184-57214397      - |   HAVANA
#>                                type     score     phase                gene_id
#>                            <factor> <numeric> <integer>            <character>
#>          ENSG00000278704.1     gene      2237      <NA>      ENSG00000278704.1
#>          ENSG00000277400.1     gene      2179      <NA>      ENSG00000277400.1
#>          ENSG00000274847.1     gene      1599      <NA>      ENSG00000274847.1
#>          ENSG00000277428.1     gene       101      <NA>      ENSG00000277428.1
#>          ENSG00000276256.1     gene      2195      <NA>      ENSG00000276256.1
#>                        ...      ...       ...       ...                    ...
#>   ENSG00000124334.17_PAR_Y     gene      2504      <NA> ENSG00000124334.17_P..
#>   ENSG00000185203.12_PAR_Y     gene      1054      <NA> ENSG00000185203.12_P..
#>    ENSG00000270726.6_PAR_Y     gene       773      <NA> ENSG00000270726.6_PA..
#>   ENSG00000182484.15_PAR_Y     gene      4618      <NA> ENSG00000182484.15_P..
#>    ENSG00000227159.8_PAR_Y     gene      1306      <NA> ENSG00000227159.8_PA..
#>                                         gene_type   gene_name       level
#>                                       <character> <character> <character>
#>          ENSG00000278704.1         protein_coding  BX004987.1           3
#>          ENSG00000277400.1         protein_coding  AC145212.2           3
#>          ENSG00000274847.1         protein_coding  AC145212.1           3
#>          ENSG00000277428.1               misc_RNA       Y_RNA           3
#>          ENSG00000276256.1         protein_coding  AC011043.1           3
#>                        ...                    ...         ...         ...
#>   ENSG00000124334.17_PAR_Y         protein_coding        IL9R           2
#>   ENSG00000185203.12_PAR_Y              antisense      WASIR1           2
#>    ENSG00000270726.6_PAR_Y   processed_transcript AJ271736.10           2
#>   ENSG00000182484.15_PAR_Y transcribed_unproces..      WASH6P           2
#>    ENSG00000227159.8_PAR_Y unprocessed_pseudogene    DDX11L16           2
#>                                     havana_gene         tag
#>                                     <character> <character>
#>          ENSG00000278704.1                 <NA>        <NA>
#>          ENSG00000277400.1                 <NA>        <NA>
#>          ENSG00000274847.1                 <NA>        <NA>
#>          ENSG00000277428.1                 <NA>        <NA>
#>          ENSG00000276256.1                 <NA>        <NA>
#>                        ...                  ...         ...
#>   ENSG00000124334.17_PAR_Y OTTHUMG00000022720.1         PAR
#>   ENSG00000185203.12_PAR_Y OTTHUMG00000022676.3         PAR
#>    ENSG00000270726.6_PAR_Y OTTHUMG00000184987.2         PAR
#>   ENSG00000182484.15_PAR_Y OTTHUMG00000022677.5         PAR
#>    ENSG00000227159.8_PAR_Y OTTHUMG00000022678.1         PAR
#>   -------
#>   seqinfo: 374 sequences from an unspecified genome; no seqlengths

colnames(colData(rse_gene_ERP110066))

#>   [1] "rail_id"                                                       
#>   [2] "run_acc"                                                       
#>   [3] "study_acc"                                                     
#>   [4] "ncbi.sample_acc"                                               
#>   [5] "ncbi.experiment_acc"                                           
#>   [6] "ncbi.submission_acc"                                           
#>   [7] "ncbi.submission_center"                                        
#>   [8] "ncbi.submission_lab"                                           
#>   [9] "ncbi.study_title"                                              
#>  [10] "ncbi.study_abstract"                                           
#>  [11] "ncbi.study_description"                                        
#>  [12] "ncbi.experiment_title"                                         
#>  [13] "ncbi.design_description"                                       
#>  [14] "ncbi.sample_description"                                       
#>  [15] "ncbi.library_name"                                             
#>  [16] "ncbi.library_strategy"                                         
#>  [17] "ncbi.library_source"                                           
#>  [18] "ncbi.library_selection"                                        
#>  [19] "ncbi.library_layout"                                           
#>  [20] "ncbi.paired_nominal_length"                                    
#>  [21] "ncbi.paired_nominal_stdev"                                     
#>  [22] "ncbi.library_construction_protocol"                            
#>  [23] "ncbi.platform_model"                                           
#>  [24] "ncbi.sample_attributes"                                        
#>  [25] "ncbi.experiment_attributes"                                    
#>  [26] "ncbi.spot_length"                                              
#>  [27] "ncbi.sample_name"                                              
#>  [28] "ncbi.sample_title"                                             
#>  [29] "ncbi.sample_bases"                                             
#>  [30] "ncbi.sample_spots"                                             
#>  [31] "ncbi.run_published"                                            
#>  [32] "ncbi.size"                                                     
#>  [33] "ncbi.run_total_bases"                                          
#>  [34] "ncbi.run_total_spots"                                          
#>  [35] "ncbi.num_reads"                                                
#>  [36] "ncbi.num_spots"                                                
#>  [37] "ncbi.read_info"                                                
#>  [38] "ncbi.run_alias"                                                
#>  [39] "ncbi.run_center_name"                                          
#>  [40] "ncbi.run_broker_name"                                          
#>  [41] "ncbi.run_center"                                               
#>  [42] "recount_project.project"                                       
#>  [43] "recount_project.organism"                                      
#>  [44] "recount_project.file_source"                                   
#>  [45] "recount_project.metadata_source"                               
#>  [46] "recount_project.date_processed"                                
#>  [47] "recount_qc.aligned_reads%.chrm"                                
#>  [48] "recount_qc.aligned_reads%.chrx"                                
#>  [49] "recount_qc.aligned_reads%.chry"                                
#>  [50] "recount_qc.bc_auc.all_reads_all_bases"                         
#>  [51] "recount_qc.bc_auc.all_reads_annotated_bases"                   
#>  [52] "recount_qc.bc_auc.unique_reads_all_bases"                      
#>  [53] "recount_qc.bc_auc.unique_reads_annotated_bases"                
#>  [54] "recount_qc.bc_auc.all_%"                                       
#>  [55] "recount_qc.bc_auc.unique_%"                                    
#>  [56] "recount_qc.bc_frag.count"                                      
#>  [57] "recount_qc.bc_frag.kallisto_count"                             
#>  [58] "recount_qc.bc_frag.kallisto_mean_length"                       
#>  [59] "recount_qc.bc_frag.mean_length"                                
#>  [60] "recount_qc.bc_frag.mode_length"                                
#>  [61] "recount_qc.bc_frag.mode_length_count"                          
#>  [62] "recount_qc.exon_fc.all_%"                                      
#>  [63] "recount_qc.exon_fc.unique_%"                                   
#>  [64] "recount_qc.exon_fc_count_all.total"                            
#>  [65] "recount_qc.exon_fc_count_all.assigned"                         
#>  [66] "recount_qc.exon_fc_count_unique.total"                         
#>  [67] "recount_qc.exon_fc_count_unique.assigned"                      
#>  [68] "recount_qc.gene_fc.all_%"                                      
#>  [69] "recount_qc.gene_fc.unique_%"                                   
#>  [70] "recount_qc.gene_fc_count_all.total"                            
#>  [71] "recount_qc.gene_fc_count_all.assigned"                         
#>  [72] "recount_qc.gene_fc_count_unique.total"                         
#>  [73] "recount_qc.gene_fc_count_unique.assigned"                      
#>  [74] "recount_qc.intron_sum"                                         
#>  [75] "recount_qc.intron_sum_%"                                       
#>  [76] "recount_qc.star.%_of_chimeric_reads"                           
#>  [77] "recount_qc.star.%_of_chimeric_reads2"                          
#>  [78] "recount_qc.star.%_of_reads_mapped_to_multiple_loci"            
#>  [79] "recount_qc.star.%_of_reads_mapped_to_multiple_loci2"           
#>  [80] "recount_qc.star.%_of_reads_mapped_to_too_many_loci"            
#>  [81] "recount_qc.star.%_of_reads_mapped_to_too_many_loci2"           
#>  [82] "recount_qc.star.%_of_reads_unmapped:_other"                    
#>  [83] "recount_qc.star.%_of_reads_unmapped:_other2"                   
#>  [84] "recount_qc.star.%_of_reads_unmapped:_too_many_mismatches"      
#>  [85] "recount_qc.star.%_of_reads_unmapped:_too_many_mismatches2"     
#>  [86] "recount_qc.star.%_of_reads_unmapped:_too_short"                
#>  [87] "recount_qc.star.%_of_reads_unmapped:_too_short2"               
#>  [88] "recount_qc.star.all_mapped_reads"                              
#>  [89] "recount_qc.star.all_mapped_reads2"                             
#>  [90] "recount_qc.star.average_input_read_length"                     
#>  [91] "recount_qc.star.average_input_read_length2"                    
#>  [92] "recount_qc.star.average_mapped_length"                         
#>  [93] "recount_qc.star.average_mapped_length2"                        
#>  [94] "recount_qc.star.deletion_average_length"                       
#>  [95] "recount_qc.star.deletion_average_length2"                      
#>  [96] "recount_qc.star.deletion_rate_per_base"                        
#>  [97] "recount_qc.star.deletion_rate_per_base2"                       
#>  [98] "recount_qc.star.insertion_average_length"                      
#>  [99] "recount_qc.star.insertion_average_length2"                     
#> [100] "recount_qc.star.insertion_rate_per_base"                       
#> [101] "recount_qc.star.insertion_rate_per_base2"                      
#> [102] "recount_qc.star.mapping_speed,_million_of_reads_per_hour"      
#> [103] "recount_qc.star.mapping_speed,_million_of_reads_per_hour2"     
#> [104] "recount_qc.star.mismatch_rate_per_base,_%"                     
#> [105] "recount_qc.star.mismatch_rate_per_base,_%2"                    
#> [106] "recount_qc.star.number_of_chimeric_reads"                      
#> [107] "recount_qc.star.number_of_chimeric_reads2"                     
#> [108] "recount_qc.star.number_of_input_reads"                         
#> [109] "recount_qc.star.number_of_input_reads2"                        
#> [110] "recount_qc.star.number_of_reads_mapped_to_multiple_loci"       
#> [111] "recount_qc.star.number_of_reads_mapped_to_multiple_loci2"      
#> [112] "recount_qc.star.number_of_reads_mapped_to_too_many_loci"       
#> [113] "recount_qc.star.number_of_reads_mapped_to_too_many_loci2"      
#> [114] "recount_qc.star.number_of_reads_unmapped:_other"               
#> [115] "recount_qc.star.number_of_reads_unmapped:_other2"              
#> [116] "recount_qc.star.number_of_reads_unmapped:_too_many_mismatches" 
#> [117] "recount_qc.star.number_of_reads_unmapped:_too_many_mismatches2"
#> [118] "recount_qc.star.number_of_reads_unmapped:_too_short"           
#> [119] "recount_qc.star.number_of_reads_unmapped:_too_short2"          
#> [120] "recount_qc.star.number_of_splices:_at/ac"                      
#> [121] "recount_qc.star.number_of_splices:_at/ac2"                     
#> [122] "recount_qc.star.number_of_splices:_annotated_(sjdb)"           
#> [123] "recount_qc.star.number_of_splices:_annotated_(sjdb)2"          
#> [124] "recount_qc.star.number_of_splices:_gc/ag"                      
#> [125] "recount_qc.star.number_of_splices:_gc/ag2"                     
#> [126] "recount_qc.star.number_of_splices:_gt/ag"                      
#> [127] "recount_qc.star.number_of_splices:_gt/ag2"                     
#> [128] "recount_qc.star.number_of_splices:_non-canonical"              
#> [129] "recount_qc.star.number_of_splices:_non-canonical2"             
#> [130] "recount_qc.star.number_of_splices:_total"                      
#> [131] "recount_qc.star.number_of_splices:_total2"                     
#> [132] "recount_qc.star.uniquely_mapped_reads_%"                       
#> [133] "recount_qc.star.uniquely_mapped_reads_%2"                      
#> [134] "recount_qc.star.uniquely_mapped_reads_number"                  
#> [135] "recount_qc.star.uniquely_mapped_reads_number2"                 
#> [136] "recount_qc.junction_count"                                     
#> [137] "recount_qc.junction_coverage"                                  
#> [138] "recount_qc.junction_avg_coverage"                              
#> [139] "BigWigURL"                                                     

## Create a second RSE object using another annotation
rse_gene_ERP110066_gencode_v29 <- create_rse(
    proj_info,
    annotation = "gencode_v29"
)

#> 2020-05-27 01:36:56 downloading and reading the metadata

#> Caching file sra.ncbi.ERP110066.MD.gz

#> Caching file sra.recount_project.ERP110066.MD.gz

#> Caching file sra.recount_qc.ERP110066.MD.gz

#> 2020-05-27 01:36:58 downloading and reading the feature information

#> Caching file human.gene_sums.G029.gtf.gz

#> adding rname 'http://snaptron.cs.jhu.edu/data/temp/recount3/human/annotations/gene_sums/human.gene_sums.G029.gtf.gz'

#> 2020-05-27 01:36:59 downloading and reading the counts: 1024 samples across 64837 features.

#> Caching file sra.gene_sums.ERP110066.G029.gz

#> adding rname 'http://snaptron.cs.jhu.edu/data/temp/recount3/human/data_sources/sra/gene_sums/66/ERP110066/sra.gene_sums.ERP110066.G029.gz'

#> 2020-05-27 01:37:04 construcing the RangedSummarizedExperiment (rse) object

## Doesn't work right now
# rse_gene_ERP110066_refseq <- create_rse(
#     proj_info,
#     annotation = "refseq"
# )

## Create a RSE object at the exon level
if (FALSE) {
rse_exon_ERP110066 <- create_rse(
    proj_info,
    type = "exon"
)

## Explore the resulting RSE exon object
rse_exon_ERP110066

dim(rse_exon_ERP110066)
rowRanges(rse_exon_ERP110066)
colData(rse_exon_ERP110066)
colnames(colData(rse_exon_ERP110066))
}