Internal function for creating a recount3 RangedSummarizedExperiment object

This function is used internally by create_rse() to construct a recount3 RangedSummarizedExperiment-class object that contains the base-pair coverage counts at the gene or exon feature level for a given annotation.

create_rse_manual(
  project,
  project_home = project_home_available(organism = organism, recount3_url =
    recount3_url, bfc = bfc),
  type = c("gene", "exon", "jxn"),
  organism = c("human", "mouse"),
  annotation = annotation_options(organism),
  bfc = BiocFileCache::BiocFileCache(),
  recount3_url = "http://snaptron.cs.jhu.edu/data/temp/recount3"
)

Arguments

project
A character(1) with the ID for a given study.

project_home
A character(1) with the home directory for the project. You can find these using project_home_available().

type
A character(1) specifying whether you want to access gene, exon, or exon-exon junction counts.

organism
A character(1) specifying which organism you want to download data from. Supported options are "human" or "mouse".

annotation
A character(1) specifying which annotation you want to download. Only used when type is either gene or exon.

bfc
A BiocFileCache-class object where the files will be cached to.

recount3_url
A character(1) specifying the home URL for recount3 or a local directory where you have mirrored recount3.

Value

A RangedSummarizedExperiment-class object.

References

https://doi.org/10.12688/f1000research.12223.1 for details on the base-pair coverage counts used in recount2 and recount3.

See also

Examples

rse_gene_ERP110066_manual <- create_rse_manual(
    "ERP110066",
    "data_sources/sra"
)

#> 2020-06-23 14:26:05 downloading and reading the metadata

#> Caching file sra.sra.ERP110066.MD.gz

#> Caching file sra.recount_project.ERP110066.MD.gz

#> Caching file sra.recount_qc.ERP110066.MD.gz

#> 2020-06-23 14:26:07 downloading and reading the feature information

#> Caching file human.gene_sums.G026.gtf.gz

#> 2020-06-23 14:26:08 downloading and reading the counts: 1024 samples across 63856 features.

#> Caching file sra.gene_sums.ERP110066.G026.gz

#> 2020-06-23 14:26:11 construcing the RangedSummarizedExperiment (rse) object

rse_gene_ERP110066_manual

#> class: RangedSummarizedExperiment 
#> dim: 63856 1024 
#> metadata(0):
#> assays(1): counts
#> rownames(63856): ENSG00000278704.1 ENSG00000277400.1 ...
#>   ENSG00000182484.15_PAR_Y ENSG00000227159.8_PAR_Y
#> rowData names(10): source type ... havana_gene tag
#> colnames(1024): ERR2713604 ERR2713605 ... ERR2713998 ERR2713999
#> colData names(139): rail_id run_acc ...
#>   recount_qc.junction_avg_coverage BigWigURL

rse_gene_ERP110066_collection_manual <- create_rse_manual(
    "ERP110066",
    "collections/geuvadis_smartseq"
)

#> 2020-06-23 14:26:12 downloading and reading the metadata

#> Caching file geuvadis_smartseq.recount_project.gz

#> adding rname 'http://snaptron.cs.jhu.edu/data/temp/recount3/human/collections/geuvadis_smartseq/metadata/geuvadis_smartseq.recount_project.gz'

#> Caching file sra.sra.ERP110066.MD.gz

#> Caching file sra.recount_project.ERP110066.MD.gz

#> Caching file sra.recount_qc.ERP110066.MD.gz

#> Caching file geuvadis_smartseq.custom.gz

#> adding rname 'http://snaptron.cs.jhu.edu/data/temp/recount3/human/collections/geuvadis_smartseq/metadata/geuvadis_smartseq.custom.gz'

#> 2020-06-23 14:26:15 downloading and reading the feature information

#> Caching file human.gene_sums.G026.gtf.gz

#> 2020-06-23 14:26:15 downloading and reading the counts: 1 sample across 63856 features.

#> Caching file sra.gene_sums.ERP110066.G026.gz

#> 2020-06-23 14:26:18 construcing the RangedSummarizedExperiment (rse) object

rse_gene_ERP110066_collection_manual

#> class: RangedSummarizedExperiment 
#> dim: 63856 1 
#> metadata(0):
#> assays(1): counts
#> rownames(63856): ENSG00000278704.1 ENSG00000277400.1 ...
#>   ENSG00000182484.15_PAR_Y ENSG00000227159.8_PAR_Y
#> rowData names(10): source type ... havana_gene tag
#> colnames(1): ERR2713106
#> colData names(140): rail_id run_acc ... custom.sequencing_type
#>   BigWigURL

rse_exon_ERP110066_collection_manual <- create_rse_manual(
    "ERP110066",
    "collections/geuvadis_smartseq",
    type = "exon"
)

#> 2020-06-23 14:26:18 downloading and reading the metadata

#> Caching file geuvadis_smartseq.recount_project.gz

#> Caching file sra.sra.ERP110066.MD.gz

#> Caching file sra.recount_project.ERP110066.MD.gz

#> Caching file sra.recount_qc.ERP110066.MD.gz

#> Caching file geuvadis_smartseq.custom.gz

#> 2020-06-23 14:26:21 downloading and reading the feature information

#> Caching file human.exon_sums.G026.gtf.gz

#> adding rname 'http://snaptron.cs.jhu.edu/data/temp/recount3/human/annotations/exon_sums/human.exon_sums.G026.gtf.gz'

#> 2020-06-23 14:26:43 downloading and reading the counts: 1 sample across 1299686 features.

#> Caching file sra.exon_sums.ERP110066.G026.gz

#> adding rname 'http://snaptron.cs.jhu.edu/data/temp/recount3/human/data_sources/sra/exon_sums/66/ERP110066/sra.exon_sums.ERP110066.G026.gz'

#> 2020-06-23 14:27:44 construcing the RangedSummarizedExperiment (rse) object

rse_exon_ERP110066_collection_manual

#> class: RangedSummarizedExperiment 
#> dim: 1299686 1 
#> metadata(0):
#> assays(1): counts
#> rownames(1299686): GL000009.2|56140|58376|- GL000194.1|53594|54832|-
#>   ... chrY|57213880|57213964|- chrY|57214350|57214397|-
#> rowData names(21): source type ... ont ccdsid
#> colnames(1): ERR2713106
#> colData names(140): rail_id run_acc ... custom.sequencing_type
#>   BigWigURL

system.time(rse_jxn_ERP110066_collection_manual <- create_rse_manual(
    "ERP110066",
    "collections/geuvadis_smartseq",
    type = "jxn"
))

#> 2020-06-23 14:27:46 downloading and reading the metadata

#> Caching file geuvadis_smartseq.recount_project.gz

#> Caching file sra.sra.ERP110066.MD.gz

#> Caching file sra.recount_project.ERP110066.MD.gz

#> Caching file sra.recount_qc.ERP110066.MD.gz

#> Caching file geuvadis_smartseq.custom.gz

#> 2020-06-23 14:27:49 downloading and reading the feature information

#> Caching file sra.junctions.ERP110066.RR.gz

#> adding rname 'http://snaptron.cs.jhu.edu/data/temp/recount3/human/data_sources/sra/junctions/66/ERP110066/sra.junctions.ERP110066.RR.gz'

#> 2020-06-23 14:27:55 downloading and reading the counts: 1 sample across 585473 features.

#> Caching file sra.junctions.ERP110066.MM.gz

#> adding rname 'http://snaptron.cs.jhu.edu/data/temp/recount3/human/data_sources/sra/junctions/66/ERP110066/sra.junctions.ERP110066.MM.gz'

#> Caching file sra.sra.ERP110066.MD.gz

#> Caching file sra.recount_project.ERP110066.MD.gz

#> Caching file sra.recount_qc.ERP110066.MD.gz

#> 2020-06-23 14:28:20 construcing the RangedSummarizedExperiment (rse) object

#>    user  system elapsed 
#>  31.708   1.978  34.533 

rse_jxn_ERP110066_collection_manual

#> class: RangedSummarizedExperiment 
#> dim: 585473 1 
#> metadata(0):
#> assays(1): counts
#> rownames: NULL
#> rowData names(7): snaptron_id length ... left_annotated right_annotated
#> colnames(1): ERR2713106
#> colData names(140): rail_id run_acc ... custom.sequencing_type
#>   BigWigURL

if (FALSE) {
project <- "ERP110066"
project_home <- "collections/geuvadis_smartseq"
type <- "jxn"
organism <- "human"
annotation <- "gencode_v26"
bfc <- BiocFileCache::BiocFileCache()
recount3_url <- "http://snaptron.cs.jhu.edu/data/temp/recount3"
}