Read the metadata files

This function reads in the recount3 metadata files into R. You can first locate the files using file_locate_url() then download it to your computer using file_retrieve().

read_metadata(metadata_files)

Arguments

metadata_files
A character() with the local path to recount3 metadata files.

Value

A data.frame() with all lower case column names for the sample metadata.

See also

Examples

## Download the metadata files for project ERP110066
url_ERP110066_meta <- file_locate_url(
    "ERP110066",
    "data_sources/sra"
)
local_ERP110066_meta <- file_retrieve(
    url = url_ERP110066_meta
)

#> Caching file sra.sra.ERP110066.MD.gz

#> Caching file sra.recount_project.ERP110066.MD.gz

#> Caching file sra.recount_qc.ERP110066.MD.gz

## Read the metadata
ERP110066_meta <- read_metadata(local_ERP110066_meta)
dim(ERP110066_meta)

#> [1] 1024  138

colnames(ERP110066_meta)

#>   [1] "rail_id"                                                       
#>   [2] "run_acc"                                                       
#>   [3] "study_acc"                                                     
#>   [4] "sra.sample_acc"                                                
#>   [5] "sra.experiment_acc"                                            
#>   [6] "sra.submission_acc"                                            
#>   [7] "sra.submission_center"                                         
#>   [8] "sra.submission_lab"                                            
#>   [9] "sra.study_title"                                               
#>  [10] "sra.study_abstract"                                            
#>  [11] "sra.study_description"                                         
#>  [12] "sra.experiment_title"                                          
#>  [13] "sra.design_description"                                        
#>  [14] "sra.sample_description"                                        
#>  [15] "sra.library_name"                                              
#>  [16] "sra.library_strategy"                                          
#>  [17] "sra.library_source"                                            
#>  [18] "sra.library_selection"                                         
#>  [19] "sra.library_layout"                                            
#>  [20] "sra.paired_nominal_length"                                     
#>  [21] "sra.paired_nominal_stdev"                                      
#>  [22] "sra.library_construction_protocol"                             
#>  [23] "sra.platform_model"                                            
#>  [24] "sra.sample_attributes"                                         
#>  [25] "sra.experiment_attributes"                                     
#>  [26] "sra.spot_length"                                               
#>  [27] "sra.sample_name"                                               
#>  [28] "sra.sample_title"                                              
#>  [29] "sra.sample_bases"                                              
#>  [30] "sra.sample_spots"                                              
#>  [31] "sra.run_published"                                             
#>  [32] "sra.size"                                                      
#>  [33] "sra.run_total_bases"                                           
#>  [34] "sra.run_total_spots"                                           
#>  [35] "sra.num_reads"                                                 
#>  [36] "sra.num_spots"                                                 
#>  [37] "sra.read_info"                                                 
#>  [38] "sra.run_alias"                                                 
#>  [39] "sra.run_center_name"                                           
#>  [40] "sra.run_broker_name"                                           
#>  [41] "sra.run_center"                                                
#>  [42] "recount_project.project"                                       
#>  [43] "recount_project.organism"                                      
#>  [44] "recount_project.file_source"                                   
#>  [45] "recount_project.metadata_source"                               
#>  [46] "recount_project.date_processed"                                
#>  [47] "recount_qc.aligned_reads%.chrm"                                
#>  [48] "recount_qc.aligned_reads%.chrx"                                
#>  [49] "recount_qc.aligned_reads%.chry"                                
#>  [50] "recount_qc.bc_auc.all_reads_all_bases"                         
#>  [51] "recount_qc.bc_auc.all_reads_annotated_bases"                   
#>  [52] "recount_qc.bc_auc.unique_reads_all_bases"                      
#>  [53] "recount_qc.bc_auc.unique_reads_annotated_bases"                
#>  [54] "recount_qc.bc_auc.all_%"                                       
#>  [55] "recount_qc.bc_auc.unique_%"                                    
#>  [56] "recount_qc.bc_frag.count"                                      
#>  [57] "recount_qc.bc_frag.kallisto_count"                             
#>  [58] "recount_qc.bc_frag.kallisto_mean_length"                       
#>  [59] "recount_qc.bc_frag.mean_length"                                
#>  [60] "recount_qc.bc_frag.mode_length"                                
#>  [61] "recount_qc.bc_frag.mode_length_count"                          
#>  [62] "recount_qc.exon_fc.all_%"                                      
#>  [63] "recount_qc.exon_fc.unique_%"                                   
#>  [64] "recount_qc.exon_fc_count_all.total"                            
#>  [65] "recount_qc.exon_fc_count_all.assigned"                         
#>  [66] "recount_qc.exon_fc_count_unique.total"                         
#>  [67] "recount_qc.exon_fc_count_unique.assigned"                      
#>  [68] "recount_qc.gene_fc.all_%"                                      
#>  [69] "recount_qc.gene_fc.unique_%"                                   
#>  [70] "recount_qc.gene_fc_count_all.total"                            
#>  [71] "recount_qc.gene_fc_count_all.assigned"                         
#>  [72] "recount_qc.gene_fc_count_unique.total"                         
#>  [73] "recount_qc.gene_fc_count_unique.assigned"                      
#>  [74] "recount_qc.intron_sum"                                         
#>  [75] "recount_qc.intron_sum_%"                                       
#>  [76] "recount_qc.star.%_of_chimeric_reads"                           
#>  [77] "recount_qc.star.%_of_chimeric_reads2"                          
#>  [78] "recount_qc.star.%_of_reads_mapped_to_multiple_loci"            
#>  [79] "recount_qc.star.%_of_reads_mapped_to_multiple_loci2"           
#>  [80] "recount_qc.star.%_of_reads_mapped_to_too_many_loci"            
#>  [81] "recount_qc.star.%_of_reads_mapped_to_too_many_loci2"           
#>  [82] "recount_qc.star.%_of_reads_unmapped:_other"                    
#>  [83] "recount_qc.star.%_of_reads_unmapped:_other2"                   
#>  [84] "recount_qc.star.%_of_reads_unmapped:_too_many_mismatches"      
#>  [85] "recount_qc.star.%_of_reads_unmapped:_too_many_mismatches2"     
#>  [86] "recount_qc.star.%_of_reads_unmapped:_too_short"                
#>  [87] "recount_qc.star.%_of_reads_unmapped:_too_short2"               
#>  [88] "recount_qc.star.all_mapped_reads"                              
#>  [89] "recount_qc.star.all_mapped_reads2"                             
#>  [90] "recount_qc.star.average_input_read_length"                     
#>  [91] "recount_qc.star.average_input_read_length2"                    
#>  [92] "recount_qc.star.average_mapped_length"                         
#>  [93] "recount_qc.star.average_mapped_length2"                        
#>  [94] "recount_qc.star.deletion_average_length"                       
#>  [95] "recount_qc.star.deletion_average_length2"                      
#>  [96] "recount_qc.star.deletion_rate_per_base"                        
#>  [97] "recount_qc.star.deletion_rate_per_base2"                       
#>  [98] "recount_qc.star.insertion_average_length"                      
#>  [99] "recount_qc.star.insertion_average_length2"                     
#> [100] "recount_qc.star.insertion_rate_per_base"                       
#> [101] "recount_qc.star.insertion_rate_per_base2"                      
#> [102] "recount_qc.star.mapping_speed,_million_of_reads_per_hour"      
#> [103] "recount_qc.star.mapping_speed,_million_of_reads_per_hour2"     
#> [104] "recount_qc.star.mismatch_rate_per_base,_%"                     
#> [105] "recount_qc.star.mismatch_rate_per_base,_%2"                    
#> [106] "recount_qc.star.number_of_chimeric_reads"                      
#> [107] "recount_qc.star.number_of_chimeric_reads2"                     
#> [108] "recount_qc.star.number_of_input_reads"                         
#> [109] "recount_qc.star.number_of_input_reads2"                        
#> [110] "recount_qc.star.number_of_reads_mapped_to_multiple_loci"       
#> [111] "recount_qc.star.number_of_reads_mapped_to_multiple_loci2"      
#> [112] "recount_qc.star.number_of_reads_mapped_to_too_many_loci"       
#> [113] "recount_qc.star.number_of_reads_mapped_to_too_many_loci2"      
#> [114] "recount_qc.star.number_of_reads_unmapped:_other"               
#> [115] "recount_qc.star.number_of_reads_unmapped:_other2"              
#> [116] "recount_qc.star.number_of_reads_unmapped:_too_many_mismatches" 
#> [117] "recount_qc.star.number_of_reads_unmapped:_too_many_mismatches2"
#> [118] "recount_qc.star.number_of_reads_unmapped:_too_short"           
#> [119] "recount_qc.star.number_of_reads_unmapped:_too_short2"          
#> [120] "recount_qc.star.number_of_splices:_at/ac"                      
#> [121] "recount_qc.star.number_of_splices:_at/ac2"                     
#> [122] "recount_qc.star.number_of_splices:_annotated_(sjdb)"           
#> [123] "recount_qc.star.number_of_splices:_annotated_(sjdb)2"          
#> [124] "recount_qc.star.number_of_splices:_gc/ag"                      
#> [125] "recount_qc.star.number_of_splices:_gc/ag2"                     
#> [126] "recount_qc.star.number_of_splices:_gt/ag"                      
#> [127] "recount_qc.star.number_of_splices:_gt/ag2"                     
#> [128] "recount_qc.star.number_of_splices:_non-canonical"              
#> [129] "recount_qc.star.number_of_splices:_non-canonical2"             
#> [130] "recount_qc.star.number_of_splices:_total"                      
#> [131] "recount_qc.star.number_of_splices:_total2"                     
#> [132] "recount_qc.star.uniquely_mapped_reads_%"                       
#> [133] "recount_qc.star.uniquely_mapped_reads_%2"                      
#> [134] "recount_qc.star.uniquely_mapped_reads_number"                  
#> [135] "recount_qc.star.uniquely_mapped_reads_number2"                 
#> [136] "recount_qc.junction_count"                                     
#> [137] "recount_qc.junction_coverage"                                  
#> [138] "recount_qc.junction_avg_coverage"                              

## Read the metadata files for a project in a collection
ERP110066_collection_meta <- read_metadata(
    metadata_files = file_retrieve(
        file_locate_url("ERP110066", "collections/geuvadis_smartseq")
    )
)

#> Caching file geuvadis_smartseq.recount_project.gz

#> Caching file sra.sra.ERP110066.MD.gz

#> Caching file sra.recount_project.ERP110066.MD.gz

#> Caching file sra.recount_qc.ERP110066.MD.gz

#> Caching file geuvadis_smartseq.custom.gz

dim(ERP110066_collection_meta)

#> [1]   1 139

colnames(ERP110066_collection_meta)

#>   [1] "rail_id"                                                       
#>   [2] "run_acc"                                                       
#>   [3] "study_acc"                                                     
#>   [4] "sra.sample_acc"                                                
#>   [5] "sra.experiment_acc"                                            
#>   [6] "sra.submission_acc"                                            
#>   [7] "sra.submission_center"                                         
#>   [8] "sra.submission_lab"                                            
#>   [9] "sra.study_title"                                               
#>  [10] "sra.study_abstract"                                            
#>  [11] "sra.study_description"                                         
#>  [12] "sra.experiment_title"                                          
#>  [13] "sra.design_description"                                        
#>  [14] "sra.sample_description"                                        
#>  [15] "sra.library_name"                                              
#>  [16] "sra.library_strategy"                                          
#>  [17] "sra.library_source"                                            
#>  [18] "sra.library_selection"                                         
#>  [19] "sra.library_layout"                                            
#>  [20] "sra.paired_nominal_length"                                     
#>  [21] "sra.paired_nominal_stdev"                                      
#>  [22] "sra.library_construction_protocol"                             
#>  [23] "sra.platform_model"                                            
#>  [24] "sra.sample_attributes"                                         
#>  [25] "sra.experiment_attributes"                                     
#>  [26] "sra.spot_length"                                               
#>  [27] "sra.sample_name"                                               
#>  [28] "sra.sample_title"                                              
#>  [29] "sra.sample_bases"                                              
#>  [30] "sra.sample_spots"                                              
#>  [31] "sra.run_published"                                             
#>  [32] "sra.size"                                                      
#>  [33] "sra.run_total_bases"                                           
#>  [34] "sra.run_total_spots"                                           
#>  [35] "sra.num_reads"                                                 
#>  [36] "sra.num_spots"                                                 
#>  [37] "sra.read_info"                                                 
#>  [38] "sra.run_alias"                                                 
#>  [39] "sra.run_center_name"                                           
#>  [40] "sra.run_broker_name"                                           
#>  [41] "sra.run_center"                                                
#>  [42] "recount_project.project"                                       
#>  [43] "recount_project.organism"                                      
#>  [44] "recount_project.file_source"                                   
#>  [45] "recount_project.metadata_source"                               
#>  [46] "recount_project.date_processed"                                
#>  [47] "recount_qc.aligned_reads%.chrm"                                
#>  [48] "recount_qc.aligned_reads%.chrx"                                
#>  [49] "recount_qc.aligned_reads%.chry"                                
#>  [50] "recount_qc.bc_auc.all_reads_all_bases"                         
#>  [51] "recount_qc.bc_auc.all_reads_annotated_bases"                   
#>  [52] "recount_qc.bc_auc.unique_reads_all_bases"                      
#>  [53] "recount_qc.bc_auc.unique_reads_annotated_bases"                
#>  [54] "recount_qc.bc_auc.all_%"                                       
#>  [55] "recount_qc.bc_auc.unique_%"                                    
#>  [56] "recount_qc.bc_frag.count"                                      
#>  [57] "recount_qc.bc_frag.kallisto_count"                             
#>  [58] "recount_qc.bc_frag.kallisto_mean_length"                       
#>  [59] "recount_qc.bc_frag.mean_length"                                
#>  [60] "recount_qc.bc_frag.mode_length"                                
#>  [61] "recount_qc.bc_frag.mode_length_count"                          
#>  [62] "recount_qc.exon_fc.all_%"                                      
#>  [63] "recount_qc.exon_fc.unique_%"                                   
#>  [64] "recount_qc.exon_fc_count_all.total"                            
#>  [65] "recount_qc.exon_fc_count_all.assigned"                         
#>  [66] "recount_qc.exon_fc_count_unique.total"                         
#>  [67] "recount_qc.exon_fc_count_unique.assigned"                      
#>  [68] "recount_qc.gene_fc.all_%"                                      
#>  [69] "recount_qc.gene_fc.unique_%"                                   
#>  [70] "recount_qc.gene_fc_count_all.total"                            
#>  [71] "recount_qc.gene_fc_count_all.assigned"                         
#>  [72] "recount_qc.gene_fc_count_unique.total"                         
#>  [73] "recount_qc.gene_fc_count_unique.assigned"                      
#>  [74] "recount_qc.intron_sum"                                         
#>  [75] "recount_qc.intron_sum_%"                                       
#>  [76] "recount_qc.star.%_of_chimeric_reads"                           
#>  [77] "recount_qc.star.%_of_chimeric_reads2"                          
#>  [78] "recount_qc.star.%_of_reads_mapped_to_multiple_loci"            
#>  [79] "recount_qc.star.%_of_reads_mapped_to_multiple_loci2"           
#>  [80] "recount_qc.star.%_of_reads_mapped_to_too_many_loci"            
#>  [81] "recount_qc.star.%_of_reads_mapped_to_too_many_loci2"           
#>  [82] "recount_qc.star.%_of_reads_unmapped:_other"                    
#>  [83] "recount_qc.star.%_of_reads_unmapped:_other2"                   
#>  [84] "recount_qc.star.%_of_reads_unmapped:_too_many_mismatches"      
#>  [85] "recount_qc.star.%_of_reads_unmapped:_too_many_mismatches2"     
#>  [86] "recount_qc.star.%_of_reads_unmapped:_too_short"                
#>  [87] "recount_qc.star.%_of_reads_unmapped:_too_short2"               
#>  [88] "recount_qc.star.all_mapped_reads"                              
#>  [89] "recount_qc.star.all_mapped_reads2"                             
#>  [90] "recount_qc.star.average_input_read_length"                     
#>  [91] "recount_qc.star.average_input_read_length2"                    
#>  [92] "recount_qc.star.average_mapped_length"                         
#>  [93] "recount_qc.star.average_mapped_length2"                        
#>  [94] "recount_qc.star.deletion_average_length"                       
#>  [95] "recount_qc.star.deletion_average_length2"                      
#>  [96] "recount_qc.star.deletion_rate_per_base"                        
#>  [97] "recount_qc.star.deletion_rate_per_base2"                       
#>  [98] "recount_qc.star.insertion_average_length"                      
#>  [99] "recount_qc.star.insertion_average_length2"                     
#> [100] "recount_qc.star.insertion_rate_per_base"                       
#> [101] "recount_qc.star.insertion_rate_per_base2"                      
#> [102] "recount_qc.star.mapping_speed,_million_of_reads_per_hour"      
#> [103] "recount_qc.star.mapping_speed,_million_of_reads_per_hour2"     
#> [104] "recount_qc.star.mismatch_rate_per_base,_%"                     
#> [105] "recount_qc.star.mismatch_rate_per_base,_%2"                    
#> [106] "recount_qc.star.number_of_chimeric_reads"                      
#> [107] "recount_qc.star.number_of_chimeric_reads2"                     
#> [108] "recount_qc.star.number_of_input_reads"                         
#> [109] "recount_qc.star.number_of_input_reads2"                        
#> [110] "recount_qc.star.number_of_reads_mapped_to_multiple_loci"       
#> [111] "recount_qc.star.number_of_reads_mapped_to_multiple_loci2"      
#> [112] "recount_qc.star.number_of_reads_mapped_to_too_many_loci"       
#> [113] "recount_qc.star.number_of_reads_mapped_to_too_many_loci2"      
#> [114] "recount_qc.star.number_of_reads_unmapped:_other"               
#> [115] "recount_qc.star.number_of_reads_unmapped:_other2"              
#> [116] "recount_qc.star.number_of_reads_unmapped:_too_many_mismatches" 
#> [117] "recount_qc.star.number_of_reads_unmapped:_too_many_mismatches2"
#> [118] "recount_qc.star.number_of_reads_unmapped:_too_short"           
#> [119] "recount_qc.star.number_of_reads_unmapped:_too_short2"          
#> [120] "recount_qc.star.number_of_splices:_at/ac"                      
#> [121] "recount_qc.star.number_of_splices:_at/ac2"                     
#> [122] "recount_qc.star.number_of_splices:_annotated_(sjdb)"           
#> [123] "recount_qc.star.number_of_splices:_annotated_(sjdb)2"          
#> [124] "recount_qc.star.number_of_splices:_gc/ag"                      
#> [125] "recount_qc.star.number_of_splices:_gc/ag2"                     
#> [126] "recount_qc.star.number_of_splices:_gt/ag"                      
#> [127] "recount_qc.star.number_of_splices:_gt/ag2"                     
#> [128] "recount_qc.star.number_of_splices:_non-canonical"              
#> [129] "recount_qc.star.number_of_splices:_non-canonical2"             
#> [130] "recount_qc.star.number_of_splices:_total"                      
#> [131] "recount_qc.star.number_of_splices:_total2"                     
#> [132] "recount_qc.star.uniquely_mapped_reads_%"                       
#> [133] "recount_qc.star.uniquely_mapped_reads_%2"                      
#> [134] "recount_qc.star.uniquely_mapped_reads_number"                  
#> [135] "recount_qc.star.uniquely_mapped_reads_number2"                 
#> [136] "recount_qc.junction_count"                                     
#> [137] "recount_qc.junction_coverage"                                  
#> [138] "recount_qc.junction_avg_coverage"                              
#> [139] "custom.sequencing_type"                                        

if (FALSE) {
## Locate and read the GTEx bladder metadata
gtex_bladder_meta <- read_metadata(
    file_retrieve(
        file_locate_url("BLADDER", "data_sources/gtex")
    )
)

head(gtex_bladder_meta)
}