Read the metadata files

This function reads in the recount3 metadata files into R. You can first locate the files using file_locate_url() then download it to your computer using file_retrieve().

read_metadata(metadata_files)

Arguments

metadata_files
A character() with the local path to recount3 metadata files.

Value

A data.frame() with all lower case column names for the sample metadata.

See also

Examples

## Download the metadata files for project ERP110066
url_ERP110066_meta <- file_locate_url(
    "ERP110066",
    "data_sources/sra"
)
local_ERP110066_meta <- file_retrieve(
    url = url_ERP110066_meta
)

#> 2020-09-28 02:43:51 caching file sra.sra.ERP110066.MD.gz.

#> adding rname 'http://idies.jhu.edu/recount3/data/human/data_sources/sra/metadata/66/ERP110066/sra.sra.ERP110066.MD.gz'

#> 2020-09-28 02:43:52 caching file sra.recount_project.ERP110066.MD.gz.

#> adding rname 'http://idies.jhu.edu/recount3/data/human/data_sources/sra/metadata/66/ERP110066/sra.recount_project.ERP110066.MD.gz'

#> 2020-09-28 02:43:53 caching file sra.recount_qc.ERP110066.MD.gz.

#> adding rname 'http://idies.jhu.edu/recount3/data/human/data_sources/sra/metadata/66/ERP110066/sra.recount_qc.ERP110066.MD.gz'

#> 2020-09-28 02:43:53 caching file sra.recount_seq_qc.ERP110066.MD.gz.

#> adding rname 'http://idies.jhu.edu/recount3/data/human/data_sources/sra/metadata/66/ERP110066/sra.recount_seq_qc.ERP110066.MD.gz'

#> 2020-09-28 02:43:53 caching file sra.recount_pred.ERP110066.MD.gz.

#> adding rname 'http://idies.jhu.edu/recount3/data/human/data_sources/sra/metadata/66/ERP110066/sra.recount_pred.ERP110066.MD.gz'

## Read the metadata
ERP110066_meta <- read_metadata(local_ERP110066_meta)

#> Warning: Replacing study_acc by study in sra.sra.ERP110066.MD.gz

#> Warning: Replacing run_acc by external_id in sra.sra.ERP110066.MD.gz

#> Warning: Replacing study_acc by study in sra.recount_qc.ERP110066.MD.gz

#> Warning: Replacing run_acc by external_id in sra.recount_qc.ERP110066.MD.gz

dim(ERP110066_meta)

#> [1] 1024  172

colnames(ERP110066_meta)

#>   [1] "rail_id"                                                           
#>   [2] "external_id"                                                       
#>   [3] "study"                                                             
#>   [4] "sra.sample_acc.x"                                                  
#>   [5] "sra.experiment_acc"                                                
#>   [6] "sra.submission_acc"                                                
#>   [7] "sra.submission_center"                                             
#>   [8] "sra.submission_lab"                                                
#>   [9] "sra.study_title"                                                   
#>  [10] "sra.study_abstract"                                                
#>  [11] "sra.study_description"                                             
#>  [12] "sra.experiment_title"                                              
#>  [13] "sra.design_description"                                            
#>  [14] "sra.sample_description"                                            
#>  [15] "sra.library_name"                                                  
#>  [16] "sra.library_strategy"                                              
#>  [17] "sra.library_source"                                                
#>  [18] "sra.library_selection"                                             
#>  [19] "sra.library_layout"                                                
#>  [20] "sra.paired_nominal_length"                                         
#>  [21] "sra.paired_nominal_stdev"                                          
#>  [22] "sra.library_construction_protocol"                                 
#>  [23] "sra.platform_model"                                                
#>  [24] "sra.sample_attributes"                                             
#>  [25] "sra.experiment_attributes"                                         
#>  [26] "sra.spot_length"                                                   
#>  [27] "sra.sample_name"                                                   
#>  [28] "sra.sample_title"                                                  
#>  [29] "sra.sample_bases"                                                  
#>  [30] "sra.sample_spots"                                                  
#>  [31] "sra.run_published"                                                 
#>  [32] "sra.size"                                                          
#>  [33] "sra.run_total_bases"                                               
#>  [34] "sra.run_total_spots"                                               
#>  [35] "sra.num_reads"                                                     
#>  [36] "sra.num_spots"                                                     
#>  [37] "sra.read_info"                                                     
#>  [38] "sra.run_alias"                                                     
#>  [39] "sra.run_center_name"                                               
#>  [40] "sra.run_broker_name"                                               
#>  [41] "sra.run_center"                                                    
#>  [42] "recount_project.project"                                           
#>  [43] "recount_project.organism"                                          
#>  [44] "recount_project.file_source"                                       
#>  [45] "recount_project.metadata_source"                                   
#>  [46] "recount_project.date_processed"                                    
#>  [47] "recount_qc.aligned_reads%.chrm"                                    
#>  [48] "recount_qc.aligned_reads%.chrx"                                    
#>  [49] "recount_qc.aligned_reads%.chry"                                    
#>  [50] "recount_qc.bc_auc.all_reads_all_bases"                             
#>  [51] "recount_qc.bc_auc.all_reads_annotated_bases"                       
#>  [52] "recount_qc.bc_auc.unique_reads_all_bases"                          
#>  [53] "recount_qc.bc_auc.unique_reads_annotated_bases"                    
#>  [54] "recount_qc.bc_auc.all_%"                                           
#>  [55] "recount_qc.bc_auc.unique_%"                                        
#>  [56] "recount_qc.bc_frag.count"                                          
#>  [57] "recount_qc.bc_frag.kallisto_count"                                 
#>  [58] "recount_qc.bc_frag.kallisto_mean_length"                           
#>  [59] "recount_qc.bc_frag.mean_length"                                    
#>  [60] "recount_qc.bc_frag.mode_length"                                    
#>  [61] "recount_qc.bc_frag.mode_length_count"                              
#>  [62] "recount_qc.exon_fc.all_%"                                          
#>  [63] "recount_qc.exon_fc.unique_%"                                       
#>  [64] "recount_qc.exon_fc_count_all.total"                                
#>  [65] "recount_qc.exon_fc_count_all.assigned"                             
#>  [66] "recount_qc.exon_fc_count_unique.total"                             
#>  [67] "recount_qc.exon_fc_count_unique.assigned"                          
#>  [68] "recount_qc.gene_fc.all_%"                                          
#>  [69] "recount_qc.gene_fc.unique_%"                                       
#>  [70] "recount_qc.gene_fc_count_all.total"                                
#>  [71] "recount_qc.gene_fc_count_all.assigned"                             
#>  [72] "recount_qc.gene_fc_count_unique.total"                             
#>  [73] "recount_qc.gene_fc_count_unique.assigned"                          
#>  [74] "recount_qc.intron_sum"                                             
#>  [75] "recount_qc.intron_sum_%"                                           
#>  [76] "recount_qc.star.%_of_chimeric_reads"                               
#>  [77] "recount_qc.star.%_of_chimeric_reads2"                              
#>  [78] "recount_qc.star.%_of_reads_mapped_to_multiple_loci"                
#>  [79] "recount_qc.star.%_of_reads_mapped_to_multiple_loci2"               
#>  [80] "recount_qc.star.%_of_reads_mapped_to_too_many_loci"                
#>  [81] "recount_qc.star.%_of_reads_mapped_to_too_many_loci2"               
#>  [82] "recount_qc.star.%_of_reads_unmapped:_other"                        
#>  [83] "recount_qc.star.%_of_reads_unmapped:_other2"                       
#>  [84] "recount_qc.star.%_of_reads_unmapped:_too_many_mismatches"          
#>  [85] "recount_qc.star.%_of_reads_unmapped:_too_many_mismatches2"         
#>  [86] "recount_qc.star.%_of_reads_unmapped:_too_short"                    
#>  [87] "recount_qc.star.%_of_reads_unmapped:_too_short2"                   
#>  [88] "recount_qc.star.all_mapped_reads"                                  
#>  [89] "recount_qc.star.all_mapped_reads2"                                 
#>  [90] "recount_qc.star.average_input_read_length"                         
#>  [91] "recount_qc.star.average_input_read_length2"                        
#>  [92] "recount_qc.star.average_mapped_length"                             
#>  [93] "recount_qc.star.average_mapped_length2"                            
#>  [94] "recount_qc.star.deletion_average_length"                           
#>  [95] "recount_qc.star.deletion_average_length2"                          
#>  [96] "recount_qc.star.deletion_rate_per_base"                            
#>  [97] "recount_qc.star.deletion_rate_per_base2"                           
#>  [98] "recount_qc.star.insertion_average_length"                          
#>  [99] "recount_qc.star.insertion_average_length2"                         
#> [100] "recount_qc.star.insertion_rate_per_base"                           
#> [101] "recount_qc.star.insertion_rate_per_base2"                          
#> [102] "recount_qc.star.mapping_speed,_million_of_reads_per_hour"          
#> [103] "recount_qc.star.mapping_speed,_million_of_reads_per_hour2"         
#> [104] "recount_qc.star.mismatch_rate_per_base,_%"                         
#> [105] "recount_qc.star.mismatch_rate_per_base,_%2"                        
#> [106] "recount_qc.star.number_of_chimeric_reads"                          
#> [107] "recount_qc.star.number_of_chimeric_reads2"                         
#> [108] "recount_qc.star.number_of_input_reads"                             
#> [109] "recount_qc.star.number_of_input_reads2"                            
#> [110] "recount_qc.star.number_of_reads_mapped_to_multiple_loci"           
#> [111] "recount_qc.star.number_of_reads_mapped_to_multiple_loci2"          
#> [112] "recount_qc.star.number_of_reads_mapped_to_too_many_loci"           
#> [113] "recount_qc.star.number_of_reads_mapped_to_too_many_loci2"          
#> [114] "recount_qc.star.number_of_reads_unmapped:_other"                   
#> [115] "recount_qc.star.number_of_reads_unmapped:_other2"                  
#> [116] "recount_qc.star.number_of_reads_unmapped:_too_many_mismatches"     
#> [117] "recount_qc.star.number_of_reads_unmapped:_too_many_mismatches2"    
#> [118] "recount_qc.star.number_of_reads_unmapped:_too_short"               
#> [119] "recount_qc.star.number_of_reads_unmapped:_too_short2"              
#> [120] "recount_qc.star.number_of_splices:_at/ac"                          
#> [121] "recount_qc.star.number_of_splices:_at/ac2"                         
#> [122] "recount_qc.star.number_of_splices:_annotated_(sjdb)"               
#> [123] "recount_qc.star.number_of_splices:_annotated_(sjdb)2"              
#> [124] "recount_qc.star.number_of_splices:_gc/ag"                          
#> [125] "recount_qc.star.number_of_splices:_gc/ag2"                         
#> [126] "recount_qc.star.number_of_splices:_gt/ag"                          
#> [127] "recount_qc.star.number_of_splices:_gt/ag2"                         
#> [128] "recount_qc.star.number_of_splices:_non-canonical"                  
#> [129] "recount_qc.star.number_of_splices:_non-canonical2"                 
#> [130] "recount_qc.star.number_of_splices:_total"                          
#> [131] "recount_qc.star.number_of_splices:_total2"                         
#> [132] "recount_qc.star.uniquely_mapped_reads_%"                           
#> [133] "recount_qc.star.uniquely_mapped_reads_%2"                          
#> [134] "recount_qc.star.uniquely_mapped_reads_number"                      
#> [135] "recount_qc.star.uniquely_mapped_reads_number2"                     
#> [136] "recount_qc.junction_count"                                         
#> [137] "recount_qc.junction_coverage"                                      
#> [138] "recount_qc.junction_avg_coverage"                                  
#> [139] "recount_qc.star.number_of_input_reads_both"                        
#> [140] "recount_qc.star.all_mapped_reads_both"                             
#> [141] "recount_qc.star.number_of_chimeric_reads_both"                     
#> [142] "recount_qc.star.number_of_reads_mapped_to_multiple_loci_both"      
#> [143] "recount_qc.star.number_of_reads_mapped_to_too_many_loci_both"      
#> [144] "recount_qc.star.number_of_reads_unmapped:_other_both"              
#> [145] "recount_qc.star.number_of_reads_unmapped:_too_many_mismatches_both"
#> [146] "recount_qc.star.number_of_reads_unmapped:_too_short_both"          
#> [147] "recount_qc.star.uniquely_mapped_reads_number_both"                 
#> [148] "recount_qc.star.%_mapped_reads_both"                               
#> [149] "recount_qc.star.%_chimeric_reads_both"                             
#> [150] "recount_qc.star.%_reads_mapped_to_multiple_loci_both"              
#> [151] "recount_qc.star.%_reads_mapped_to_too_many_loci_both"              
#> [152] "recount_qc.star.%_reads_unmapped:_other_both"                      
#> [153] "recount_qc.star.%_reads_unmapped:_too_many_mismatches_both"        
#> [154] "recount_qc.star.%_reads_unmapped:_too_short_both"                  
#> [155] "recount_qc.star.uniquely_mapped_reads_%_both"                      
#> [156] "recount_seq_qc.min_len"                                            
#> [157] "recount_seq_qc.max_len"                                            
#> [158] "recount_seq_qc.avg_len"                                            
#> [159] "recount_seq_qc.#distinct_quality_values"                           
#> [160] "recount_seq_qc.#bases"                                             
#> [161] "recount_seq_qc.%a"                                                 
#> [162] "recount_seq_qc.%c"                                                 
#> [163] "recount_seq_qc.%g"                                                 
#> [164] "recount_seq_qc.%t"                                                 
#> [165] "recount_seq_qc.%n"                                                 
#> [166] "recount_seq_qc.avgq"                                               
#> [167] "recount_seq_qc.errq"                                               
#> [168] "recount_pred.sample_acc.y"                                         
#> [169] "recount_pred.curated.type"                                         
#> [170] "recount_pred.curated.tissue"                                       
#> [171] "recount_pred.pattern.predict.type"                                 
#> [172] "recount_pred.pred.type"                                            

## Read the metadata files for a project in a collection
## Note: using the test files since I can't access collections right now
## for this collection
ERP110066_collection_meta <- read_metadata(
    metadata_files = file_retrieve(
        file_locate_url(
            "ERP110066",
            "collections/geuvadis_smartseq",
            recount3_url = "http://snaptron.cs.jhu.edu/data/temp/recount3"
        )
    )
)

#> 2020-09-28 02:43:54 caching file geuvadis_smartseq.recount_project.gz.

#> 2020-09-28 02:43:55 caching file sra.sra.ERP110066.MD.gz.

#> 2020-09-28 02:43:55 caching file sra.recount_project.ERP110066.MD.gz.

#> 2020-09-28 02:43:56 caching file sra.recount_qc.ERP110066.MD.gz.

#> Warning: The 'url' <http://snaptron.cs.jhu.edu/data/temp/recount3/human/data_sources/sra/metadata/66/ERP110066/sra.recount_seq_qc.ERP110066.MD.gz> does not exist or is not available.

#> 2020-09-28 02:43:56 caching file sra.recount_pred.ERP110066.MD.gz.

#> 2020-09-28 02:43:56 caching file geuvadis_smartseq.custom.gz.

dim(ERP110066_collection_meta)

#> [1]   1 161

## New columns for this collection
colnames(ERP110066_collection_meta)[!colnames(ERP110066_collection_meta) %in% colnames(ERP110066_meta)]

#> [1] "custom.sequencing_type"

## Read the metadata for a mouse project
DRP002367_meta <- read_metadata(
    metadata_files = file_retrieve(
        file_locate_url("DRP002367", "data_sources/sra", organism = "mouse")
    )
)

#> 2020-09-28 02:43:57 caching file sra.sra.DRP002367.MD.gz.

#> 2020-09-28 02:43:57 caching file sra.recount_project.DRP002367.MD.gz.

#> 2020-09-28 02:43:57 caching file sra.recount_qc.DRP002367.MD.gz.

#> 2020-09-28 02:43:57 caching file sra.recount_seq_qc.DRP002367.MD.gz.

#> 2020-09-28 02:43:58 caching file sra.recount_pred.DRP002367.MD.gz.

#> Warning: Replacing study_acc by study in sra.sra.DRP002367.MD.gz

#> Warning: Replacing run_acc by external_id in sra.sra.DRP002367.MD.gz

#> Warning: Replacing study_acc by study in sra.recount_qc.DRP002367.MD.gz

#> Warning: Replacing run_acc by external_id in sra.recount_qc.DRP002367.MD.gz

dim(DRP002367_meta)

#> [1]   4 176

## Locate and read the GTEx bladder metadata
gtex_bladder_meta <- read_metadata(
    file_retrieve(
        file_locate_url("BLADDER", "data_sources/gtex")
    )
)

#> 2020-09-28 02:43:58 caching file gtex.gtex.BLADDER.MD.gz.

#> adding rname 'http://idies.jhu.edu/recount3/data/human/data_sources/gtex/metadata/ER/BLADDER/gtex.gtex.BLADDER.MD.gz'

#> 2020-09-28 02:43:58 caching file gtex.recount_project.BLADDER.MD.gz.

#> adding rname 'http://idies.jhu.edu/recount3/data/human/data_sources/gtex/metadata/ER/BLADDER/gtex.recount_project.BLADDER.MD.gz'

#> 2020-09-28 02:43:59 caching file gtex.recount_qc.BLADDER.MD.gz.

#> adding rname 'http://idies.jhu.edu/recount3/data/human/data_sources/gtex/metadata/ER/BLADDER/gtex.recount_qc.BLADDER.MD.gz'

#> 2020-09-28 02:43:59 caching file gtex.recount_seq_qc.BLADDER.MD.gz.

#> adding rname 'http://idies.jhu.edu/recount3/data/human/data_sources/gtex/metadata/ER/BLADDER/gtex.recount_seq_qc.BLADDER.MD.gz'

dim(gtex_bladder_meta)

#> [1]  21 197

colnames(gtex_bladder_meta)

#>   [1] "rail_id"                                                           
#>   [2] "external_id"                                                       
#>   [3] "study"                                                             
#>   [4] "gtex.run_acc"                                                      
#>   [5] "gtex.subjid"                                                       
#>   [6] "gtex.sex"                                                          
#>   [7] "gtex.age"                                                          
#>   [8] "gtex.dthhrdy"                                                      
#>   [9] "gtex.sampid"                                                       
#>  [10] "gtex.smatsscr"                                                     
#>  [11] "gtex.smcenter"                                                     
#>  [12] "gtex.smpthnts"                                                     
#>  [13] "gtex.smrin"                                                        
#>  [14] "gtex.smts"                                                         
#>  [15] "gtex.smtsd"                                                        
#>  [16] "gtex.smubrid"                                                      
#>  [17] "gtex.smtsisch"                                                     
#>  [18] "gtex.smtspax"                                                      
#>  [19] "gtex.smnabtch"                                                     
#>  [20] "gtex.smnabtcht"                                                    
#>  [21] "gtex.smnabtchd"                                                    
#>  [22] "gtex.smgebtch"                                                     
#>  [23] "gtex.smgebtchd"                                                    
#>  [24] "gtex.smgebtcht"                                                    
#>  [25] "gtex.smafrze"                                                      
#>  [26] "gtex.smgtc"                                                        
#>  [27] "gtex.sme2mprt"                                                     
#>  [28] "gtex.smchmprs"                                                     
#>  [29] "gtex.smntrart"                                                     
#>  [30] "gtex.smnumgps"                                                     
#>  [31] "gtex.smmaprt"                                                      
#>  [32] "gtex.smexncrt"                                                     
#>  [33] "gtex.sm550nrm"                                                     
#>  [34] "gtex.smgnsdtc"                                                     
#>  [35] "gtex.smunmprt"                                                     
#>  [36] "gtex.sm350nrm"                                                     
#>  [37] "gtex.smrdlgth"                                                     
#>  [38] "gtex.smmncpb"                                                      
#>  [39] "gtex.sme1mmrt"                                                     
#>  [40] "gtex.smsflgth"                                                     
#>  [41] "gtex.smestlbs"                                                     
#>  [42] "gtex.smmppd"                                                       
#>  [43] "gtex.smnterrt"                                                     
#>  [44] "gtex.smrrnanm"                                                     
#>  [45] "gtex.smrdttl"                                                      
#>  [46] "gtex.smvqcfl"                                                      
#>  [47] "gtex.smmncv"                                                       
#>  [48] "gtex.smtrscpt"                                                     
#>  [49] "gtex.smmppdpr"                                                     
#>  [50] "gtex.smcglgth"                                                     
#>  [51] "gtex.smgappct"                                                     
#>  [52] "gtex.smunpdrd"                                                     
#>  [53] "gtex.smntrnrt"                                                     
#>  [54] "gtex.smmpunrt"                                                     
#>  [55] "gtex.smexpeff"                                                     
#>  [56] "gtex.smmppdun"                                                     
#>  [57] "gtex.sme2mmrt"                                                     
#>  [58] "gtex.sme2anti"                                                     
#>  [59] "gtex.smaltalg"                                                     
#>  [60] "gtex.sme2snse"                                                     
#>  [61] "gtex.smmflgth"                                                     
#>  [62] "gtex.sme1anti"                                                     
#>  [63] "gtex.smspltrd"                                                     
#>  [64] "gtex.smbsmmrt"                                                     
#>  [65] "gtex.sme1snse"                                                     
#>  [66] "gtex.sme1pcts"                                                     
#>  [67] "gtex.smrrnart"                                                     
#>  [68] "gtex.sme1mprt"                                                     
#>  [69] "gtex.smnum5cd"                                                     
#>  [70] "gtex.smdpmprt"                                                     
#>  [71] "gtex.sme2pcts"                                                     
#>  [72] "recount_project.project"                                           
#>  [73] "recount_project.organism"                                          
#>  [74] "recount_project.file_source"                                       
#>  [75] "recount_project.metadata_source"                                   
#>  [76] "recount_project.date_processed"                                    
#>  [77] "recount_qc.aligned_reads%.chrm"                                    
#>  [78] "recount_qc.aligned_reads%.chrx"                                    
#>  [79] "recount_qc.aligned_reads%.chry"                                    
#>  [80] "recount_qc.bc_auc.all_reads_all_bases"                             
#>  [81] "recount_qc.bc_auc.all_reads_annotated_bases"                       
#>  [82] "recount_qc.bc_auc.unique_reads_all_bases"                          
#>  [83] "recount_qc.bc_auc.unique_reads_annotated_bases"                    
#>  [84] "recount_qc.bc_auc.all_%"                                           
#>  [85] "recount_qc.bc_auc.unique_%"                                        
#>  [86] "recount_qc.bc_frag.count"                                          
#>  [87] "recount_qc.bc_frag.kallisto_count"                                 
#>  [88] "recount_qc.bc_frag.kallisto_mean_length"                           
#>  [89] "recount_qc.bc_frag.mean_length"                                    
#>  [90] "recount_qc.bc_frag.mode_length"                                    
#>  [91] "recount_qc.bc_frag.mode_length_count"                              
#>  [92] "recount_qc.exon_fc.all_%"                                          
#>  [93] "recount_qc.exon_fc.unique_%"                                       
#>  [94] "recount_qc.exon_fc_count_all.total"                                
#>  [95] "recount_qc.exon_fc_count_all.assigned"                             
#>  [96] "recount_qc.exon_fc_count_unique.total"                             
#>  [97] "recount_qc.exon_fc_count_unique.assigned"                          
#>  [98] "recount_qc.gene_fc.all_%"                                          
#>  [99] "recount_qc.gene_fc.unique_%"                                       
#> [100] "recount_qc.gene_fc_count_all.total"                                
#> [101] "recount_qc.gene_fc_count_all.assigned"                             
#> [102] "recount_qc.gene_fc_count_unique.total"                             
#> [103] "recount_qc.gene_fc_count_unique.assigned"                          
#> [104] "recount_qc.intron_sum"                                             
#> [105] "recount_qc.intron_sum_%"                                           
#> [106] "recount_qc.star.%_of_chimeric_reads"                               
#> [107] "recount_qc.star.%_of_chimeric_reads2"                              
#> [108] "recount_qc.star.%_of_reads_mapped_to_multiple_loci"                
#> [109] "recount_qc.star.%_of_reads_mapped_to_multiple_loci2"               
#> [110] "recount_qc.star.%_of_reads_mapped_to_too_many_loci"                
#> [111] "recount_qc.star.%_of_reads_mapped_to_too_many_loci2"               
#> [112] "recount_qc.star.%_of_reads_unmapped:_other"                        
#> [113] "recount_qc.star.%_of_reads_unmapped:_other2"                       
#> [114] "recount_qc.star.%_of_reads_unmapped:_too_many_mismatches"          
#> [115] "recount_qc.star.%_of_reads_unmapped:_too_many_mismatches2"         
#> [116] "recount_qc.star.%_of_reads_unmapped:_too_short"                    
#> [117] "recount_qc.star.%_of_reads_unmapped:_too_short2"                   
#> [118] "recount_qc.star.all_mapped_reads"                                  
#> [119] "recount_qc.star.all_mapped_reads2"                                 
#> [120] "recount_qc.star.average_input_read_length"                         
#> [121] "recount_qc.star.average_input_read_length2"                        
#> [122] "recount_qc.star.average_mapped_length"                             
#> [123] "recount_qc.star.average_mapped_length2"                            
#> [124] "recount_qc.star.deletion_average_length"                           
#> [125] "recount_qc.star.deletion_average_length2"                          
#> [126] "recount_qc.star.deletion_rate_per_base"                            
#> [127] "recount_qc.star.deletion_rate_per_base2"                           
#> [128] "recount_qc.star.insertion_average_length"                          
#> [129] "recount_qc.star.insertion_average_length2"                         
#> [130] "recount_qc.star.insertion_rate_per_base"                           
#> [131] "recount_qc.star.insertion_rate_per_base2"                          
#> [132] "recount_qc.star.mapping_speed,_million_of_reads_per_hour"          
#> [133] "recount_qc.star.mapping_speed,_million_of_reads_per_hour2"         
#> [134] "recount_qc.star.mismatch_rate_per_base,_%"                         
#> [135] "recount_qc.star.mismatch_rate_per_base,_%2"                        
#> [136] "recount_qc.star.number_of_chimeric_reads"                          
#> [137] "recount_qc.star.number_of_chimeric_reads2"                         
#> [138] "recount_qc.star.number_of_input_reads"                             
#> [139] "recount_qc.star.number_of_input_reads2"                            
#> [140] "recount_qc.star.number_of_reads_mapped_to_multiple_loci"           
#> [141] "recount_qc.star.number_of_reads_mapped_to_multiple_loci2"          
#> [142] "recount_qc.star.number_of_reads_mapped_to_too_many_loci"           
#> [143] "recount_qc.star.number_of_reads_mapped_to_too_many_loci2"          
#> [144] "recount_qc.star.number_of_reads_unmapped:_other"                   
#> [145] "recount_qc.star.number_of_reads_unmapped:_other2"                  
#> [146] "recount_qc.star.number_of_reads_unmapped:_too_many_mismatches"     
#> [147] "recount_qc.star.number_of_reads_unmapped:_too_many_mismatches2"    
#> [148] "recount_qc.star.number_of_reads_unmapped:_too_short"               
#> [149] "recount_qc.star.number_of_reads_unmapped:_too_short2"              
#> [150] "recount_qc.star.number_of_splices:_at/ac"                          
#> [151] "recount_qc.star.number_of_splices:_at/ac2"                         
#> [152] "recount_qc.star.number_of_splices:_annotated_(sjdb)"               
#> [153] "recount_qc.star.number_of_splices:_annotated_(sjdb)2"              
#> [154] "recount_qc.star.number_of_splices:_gc/ag"                          
#> [155] "recount_qc.star.number_of_splices:_gc/ag2"                         
#> [156] "recount_qc.star.number_of_splices:_gt/ag"                          
#> [157] "recount_qc.star.number_of_splices:_gt/ag2"                         
#> [158] "recount_qc.star.number_of_splices:_non-canonical"                  
#> [159] "recount_qc.star.number_of_splices:_non-canonical2"                 
#> [160] "recount_qc.star.number_of_splices:_total"                          
#> [161] "recount_qc.star.number_of_splices:_total2"                         
#> [162] "recount_qc.star.uniquely_mapped_reads_%"                           
#> [163] "recount_qc.star.uniquely_mapped_reads_%2"                          
#> [164] "recount_qc.star.uniquely_mapped_reads_number"                      
#> [165] "recount_qc.star.uniquely_mapped_reads_number2"                     
#> [166] "recount_qc.junction_count"                                         
#> [167] "recount_qc.junction_coverage"                                      
#> [168] "recount_qc.junction_avg_coverage"                                  
#> [169] "recount_qc.star.number_of_input_reads_both"                        
#> [170] "recount_qc.star.all_mapped_reads_both"                             
#> [171] "recount_qc.star.number_of_chimeric_reads_both"                     
#> [172] "recount_qc.star.number_of_reads_mapped_to_multiple_loci_both"      
#> [173] "recount_qc.star.number_of_reads_mapped_to_too_many_loci_both"      
#> [174] "recount_qc.star.number_of_reads_unmapped:_other_both"              
#> [175] "recount_qc.star.number_of_reads_unmapped:_too_many_mismatches_both"
#> [176] "recount_qc.star.number_of_reads_unmapped:_too_short_both"          
#> [177] "recount_qc.star.uniquely_mapped_reads_number_both"                 
#> [178] "recount_qc.star.%_mapped_reads_both"                               
#> [179] "recount_qc.star.%_chimeric_reads_both"                             
#> [180] "recount_qc.star.%_reads_mapped_to_multiple_loci_both"              
#> [181] "recount_qc.star.%_reads_mapped_to_too_many_loci_both"              
#> [182] "recount_qc.star.%_reads_unmapped:_other_both"                      
#> [183] "recount_qc.star.%_reads_unmapped:_too_many_mismatches_both"        
#> [184] "recount_qc.star.%_reads_unmapped:_too_short_both"                  
#> [185] "recount_qc.star.uniquely_mapped_reads_%_both"                      
#> [186] "recount_seq_qc.min_len"                                            
#> [187] "recount_seq_qc.max_len"                                            
#> [188] "recount_seq_qc.avg_len"                                            
#> [189] "recount_seq_qc.#distinct_quality_values"                           
#> [190] "recount_seq_qc.#bases"                                             
#> [191] "recount_seq_qc.%a"                                                 
#> [192] "recount_seq_qc.%c"                                                 
#> [193] "recount_seq_qc.%g"                                                 
#> [194] "recount_seq_qc.%t"                                                 
#> [195] "recount_seq_qc.%n"                                                 
#> [196] "recount_seq_qc.avgq"                                               
#> [197] "recount_seq_qc.errq"